refactor: split CoreVM.hpp monolith into 14 focused headers
Decompose the 4,124-line CoreVM.hpp into focused headers following the
existing source directory structure (ir/, vm/, types/, transform/).
CoreVM.hpp becomes a 27-line umbrella header for backward compatibility.

Cross-layer circular dependencies resolved via forward declarations and
out-of-line definitions for methods crossing the IR/VM boundary.
