+++ src/CoreVM/CoreVM.hpp
@@ -1,4123 +1,27 @@
 // SPDX-License-Identifier: Apache-2.0
+//
+// Umbrella header — includes all CoreVM public headers.
+// Consumers that only need a subset should include specific headers instead.
+//
 #pragma once
 
+#include <CoreVM/CoreTypes.hpp>
+#include <CoreVM/Diagnostics.hpp>
+#include <CoreVM/Formatters.hpp>
 #include <CoreVM/RuntimeConfig.hpp>
+#include <CoreVM/SourceLocation.hpp>
+#include <CoreVM/TargetCodeGenerator.hpp>
 #include <CoreVM/enums.hpp>
+#include <CoreVM/ir/BasicBlock.hpp>
+#include <CoreVM/ir/IRProgram.hpp>
+#include <CoreVM/ir/Instructions.hpp>
+#include <CoreVM/ir/Value.hpp>
+#include <CoreVM/transform/Passes.hpp>
+#include <CoreVM/types/TypeDescriptor.hpp>
 #include <CoreVM/types/TypeRegistry.hpp>
 #include <CoreVM/types/TypedObject.hpp>
 #include <CoreVM/util.hpp>
-
-#include <cassert>
-#include <cstdint>
-#include <cstdio>
-#include <cstdlib>
-#include <cstring> // memset()
-#include <deque>
-#include <expected>
-#include <format>
-#include <functional> // hash<>
-#include <iostream>
-#include <list>
-#include <memory>
-#include <optional>
-#include <print>
-#include <regex>
-#include <string>
-#include <unordered_map>
-#include <unordered_set>
-#include <utility>
-#include <variant>
-#include <vector>
-
-#if defined(_WIN32) || defined(_WIN64)
-    #include <winsock2.h>
-    #include <ws2tcpip.h>
-#else
-    #include <arpa/inet.h>  // ntohl(), htonl()
-    #include <netinet/in.h> // in_addr, in6_addr
-#endif
-
-namespace CoreVM::diagnostics
-{
-class Report;
-}
-
-namespace CoreVM
-{
-
-class Instr;
-class IRBuilder;
-class IRProgram;
-class NativeCallback;
-class Value;
-class Constant;
-class BasicBlock;
-class IRFunction;
-class IRProgram;
-class IRBuilder;
-class IRBuiltinFunction;
-class AllocaInstr;
-class MatchInstr;
-class InstructionVisitor;
-class ConstantPool;
-class Program;
-class Function;
-class Runner;
-struct SourceLocation;
-struct RuntimeError;
-class Runtime;
-class NopInstr;
-class AllocaInstr;
-class StoreInstr;
-class LoadInstr;
-class CallInstr;
-class PhiNode;
-class CondBrInstr;
-class BrInstr;
-class RetInstr;
-class MatchInstr;
-class RegExpGroupInstr;
-class CastInstr;
-
-// User-defined function call instructions
-class FunctionCallInstr;
-class FunctionRetInstr;
-class TailCallInstr;
-class IndirectCallInstr;
-class IndirectTailCallInstr;
-
-// Lazy evaluation
-class FunctionRefInstr;
-class LazyForceInstr;
-
-// Object operations (for composite types)
-class ObjAllocInstr;
-class ObjRetainInstr;
-class ObjReleaseInstr;
-class ObjGetTagInstr;
-class ObjSetTagInstr;
-class ObjGetSlotInstr;
-class ObjSetSlotInstr;
-class ObjTypeIdInstr;
-class ObjIsTypeInstr;
-class VCmpEQInstr;
-class VCmpNEInstr;
-class VCmpLTInstr;
-class VCmpLEInstr;
-class VCmpGTInstr;
-class VCmpGEInstr;
-
-using Register = uint64_t; // vm
-using CoreNumber = int64_t;
-using CoreString = std::string;
-using Quota = int64_t;
-constexpr Quota NoQuota = -1;
-
-/**
- * Disassembles the @p program with @p n instructions.
- *
- * @param program pointer to the first instruction to disassemble
- * @param n       number of instructions to disassemble
- * @param indent  prefix to inject in front of every new instruction line
- * @param cp      pointer to ConstantPool for pretty-printing or @c nullptr
- *
- * @returns       disassembled program in text form.
- */
-std::string disassemble(const Instruction* program,
-                        size_t n,
-                        const std::string& indent,
-                        const ConstantPool* cp);
-
-/**
- * Disassembles a single instruction.
- *
- * @param pc      Instruction to disassemble.
- * @param ip      The instruction pointer at which position the instruction is
- *                located within the program.
- * @param sp      current stack size (depth) before executing given instruction.
- *                This value will be modified as if the instruction would have
- *                been executed.
- * @param cp      pointer to ConstantPool for pretty-printing or @c nullptr
- */
-std::string disassemble(Instruction pc, size_t ip, size_t sp, const ConstantPool* cp);
-
-std::string tos(LiteralType type);
-
-namespace diagnostics
-{
-    enum class Type;
-    std::string_view tos(Type type);
-} // namespace diagnostics
-
-bool isArrayType(LiteralType type);
-LiteralType elementTypeOf(LiteralType type);
-
-class InstructionVisitor
-{
-  public:
-    virtual ~InstructionVisitor() = default;
-
-    virtual void visit(NopInstr& instr) = 0;
-
-    // storage
-    virtual void visit(AllocaInstr& instr) = 0;
-    virtual void visit(StoreInstr& instr) = 0;
-    virtual void visit(LoadInstr& instr) = 0;
-    virtual void visit(PhiNode& instr) = 0;
-
-    // calls
-    virtual void visit(CallInstr& instr) = 0;
-
-    // user-defined function calls
-    virtual void visit(FunctionCallInstr& instr) = 0;
-    virtual void visit(FunctionRetInstr& instr) = 0;
-    virtual void visit(TailCallInstr& instr) = 0;
-
-    // terminator
-    virtual void visit(CondBrInstr& instr) = 0;
-    virtual void visit(BrInstr& instr) = 0;
-    virtual void visit(RetInstr& instr) = 0;
-    virtual void visit(MatchInstr& instr) = 0;
-
-    // regexp
-    virtual void visit(RegExpGroupInstr& instr) = 0;
-
-    // type cast
-    virtual void visit(CastInstr& instr) = 0;
-
-    // numeric
-    virtual void visit(INegInstr& instr) = 0;
-    virtual void visit(INotInstr& instr) = 0;
-    virtual void visit(IAddInstr& instr) = 0;
-    virtual void visit(ISubInstr& instr) = 0;
-    virtual void visit(IMulInstr& instr) = 0;
-    virtual void visit(IDivInstr& instr) = 0;
-    virtual void visit(IRemInstr& instr) = 0;
-    virtual void visit(IPowInstr& instr) = 0;
-    virtual void visit(IAndInstr& instr) = 0;
-    virtual void visit(IOrInstr& instr) = 0;
-    virtual void visit(IXorInstr& instr) = 0;
-    virtual void visit(IShlInstr& instr) = 0;
-    virtual void visit(IShrInstr& instr) = 0;
-    virtual void visit(ICmpEQInstr& instr) = 0;
-    virtual void visit(ICmpNEInstr& instr) = 0;
-    virtual void visit(ICmpLEInstr& instr) = 0;
-    virtual void visit(ICmpGEInstr& instr) = 0;
-    virtual void visit(ICmpLTInstr& instr) = 0;
-    virtual void visit(ICmpGTInstr& instr) = 0;
-
-    // boolean
-    virtual void visit(BNotInstr& instr) = 0;
-    virtual void visit(BAndInstr& instr) = 0;
-    virtual void visit(BOrInstr& instr) = 0;
-    virtual void visit(BXorInstr& instr) = 0;
-
-    // string
-    virtual void visit(SLenInstr& instr) = 0;
-    virtual void visit(SIsEmptyInstr& instr) = 0;
-    virtual void visit(SAddInstr& instr) = 0;
-    virtual void visit(SSubStrInstr& instr) = 0;
-    virtual void visit(SCmpEQInstr& instr) = 0;
-    virtual void visit(SCmpNEInstr& instr) = 0;
-    virtual void visit(SCmpLEInstr& instr) = 0;
-    virtual void visit(SCmpGEInstr& instr) = 0;
-    virtual void visit(SCmpLTInstr& instr) = 0;
-    virtual void visit(SCmpGTInstr& instr) = 0;
-    virtual void visit(SCmpREInstr& instr) = 0;
-    virtual void visit(SCmpBegInstr& instr) = 0;
-    virtual void visit(SCmpEndInstr& instr) = 0;
-    virtual void visit(SInInstr& instr) = 0;
-
-    // ip
-    virtual void visit(PCmpEQInstr& instr) = 0;
-    virtual void visit(PCmpNEInstr& instr) = 0;
-    virtual void visit(PInCidrInstr& instr) = 0;
-
-    // float
-    virtual void visit(FNegInstr& instr) = 0;
-    virtual void visit(FAddInstr& instr) = 0;
-    virtual void visit(FSubInstr& instr) = 0;
-    virtual void visit(FMulInstr& instr) = 0;
-    virtual void visit(FDivInstr& instr) = 0;
-    virtual void visit(FRemInstr& instr) = 0;
-    virtual void visit(FPowInstr& instr) = 0;
-    virtual void visit(FCmpEQInstr& instr) = 0;
-    virtual void visit(FCmpNEInstr& instr) = 0;
-    virtual void visit(FCmpLEInstr& instr) = 0;
-    virtual void visit(FCmpGEInstr& instr) = 0;
-    virtual void visit(FCmpLTInstr& instr) = 0;
-    virtual void visit(FCmpGTInstr& instr) = 0;
-
-    // object operations
-    virtual void visit(ObjAllocInstr& instr) = 0;
-    virtual void visit(ObjRetainInstr& instr) = 0;
-    virtual void visit(ObjReleaseInstr& instr) = 0;
-    virtual void visit(ObjGetTagInstr& instr) = 0;
-    virtual void visit(ObjSetTagInstr& instr) = 0;
-    virtual void visit(ObjGetSlotInstr& instr) = 0;
-    virtual void visit(ObjSetSlotInstr& instr) = 0;
-    virtual void visit(ObjTypeIdInstr& instr) = 0;
-    virtual void visit(ObjIsTypeInstr& instr) = 0;
-    virtual void visit(VCmpEQInstr& instr) = 0;
-    virtual void visit(VCmpNEInstr& instr) = 0;
-    virtual void visit(VCmpLTInstr& instr) = 0;
-    virtual void visit(VCmpLEInstr& instr) = 0;
-    virtual void visit(VCmpGTInstr& instr) = 0;
-    virtual void visit(VCmpGEInstr& instr) = 0;
-
-    // indirect function calls
-    virtual void visit(IndirectCallInstr& instr) = 0;
-    virtual void visit(IndirectTailCallInstr& instr) = 0;
-
-    // lazy evaluation
-    virtual void visit(FunctionRefInstr& instr) = 0;
-    virtual void visit(LazyForceInstr& instr) = 0;
-};
-
-// {{{ array types
-class CoreArray
-{
-  public:
-    [[nodiscard]] size_t size() const { return static_cast<size_t>(_base[0]); }
-
-  protected:
-    explicit CoreArray(const Register* base): _base(base) {}
-
-    [[nodiscard]] Register getRawAt(size_t i) const { return _base[1 + i]; }
-
-    [[nodiscard]] const Register* data() const { return _base + 1; }
-
-  protected:
-    const Register* _base;
-};
-
-using CoreIntArray = std::vector<CoreNumber>;
-using CoreStringArray = std::vector<CoreString>;
-using CoreIPAddrArray = std::vector<util::IPAddress>;
-using CoreCidrArray = std::vector<util::Cidr>;
-
-// }}}
-
-struct FilePos
-{ // {{{
-
-    FilePos(): FilePos { 1, 1, 0 } {}
-
-    FilePos(unsigned r, unsigned c): FilePos { r, c, 0 } {}
-
-    FilePos(unsigned r, unsigned c, unsigned o): line(r), column(c), offset(o) {}
-
-    FilePos& set(unsigned r, unsigned c, unsigned o)
-    {
-        line = r;
-        column = c;
-        offset = o;
-
-        return *this;
-    }
-
-    void advance(char ch)
-    {
-        offset++;
-        if (ch != '\n')
-        {
-            column++;
-        }
-        else
-        {
-            line++;
-            column = 1;
-        }
-    }
-
-    bool operator==(const FilePos& other) const noexcept
-    {
-        return line == other.line && column == other.column && offset == other.offset;
-    }
-
-    bool operator!=(const FilePos& other) const noexcept { return !(*this == other); }
-
-    unsigned line;
-    unsigned column;
-    unsigned offset;
-};
-
-inline size_t operator-(const FilePos& a, const FilePos& b)
-{
-    if (b.offset > a.offset)
-        return 1 + b.offset - a.offset;
-    else
-        return 1 + a.offset - b.offset;
-}
-
-/// Converts a FilePos to string (e.g. "1:5").
-std::string tos(FilePos const& pos);
-
-// }}}
-
-struct SourceLocation // {{{
-{
-    SourceLocation() = default;
-
-    SourceLocation(std::string fileName): filename(std::move(fileName)) {}
-
-    SourceLocation(std::string fileName, FilePos beg, FilePos end):
-        filename(std::move(fileName)), begin(beg), end(end)
-    {
-    }
-
-    std::string filename;
-    FilePos begin;
-    FilePos end;
-
-    SourceLocation& update(const FilePos& endPos)
-    {
-        end = endPos;
-        return *this;
-    }
-
-    SourceLocation& update(const SourceLocation& endLocation)
-    {
-        end = endLocation.end;
-        return *this;
-    }
-
-    [[nodiscard]] std::string str() const;
-    [[nodiscard]] std::string text() const;
-
-    bool operator==(const SourceLocation& other) const noexcept
-    {
-        return filename == other.filename && begin == other.begin && end == other.end;
-    }
-
-    bool operator!=(const SourceLocation& other) const noexcept { return !(*this == other); }
-}; // }}}
-
-inline SourceLocation operator-(const SourceLocation& end, const SourceLocation& beg)
-{
-    return SourceLocation(beg.filename, beg.begin, end.end);
-}
-
-// ExecutionEngine
-// VM
-class Runner
-{
-  public:
-    enum State
-    {
-        Inactive,  //!< No function running nor suspended.
-        Running,   //!< Active function is currently running.
-        Suspended, //!< Active function is currently suspended.
-    };
-
-    class QuotaExceeded: public std::runtime_error
-    {
-      public:
-        QuotaExceeded(): std::runtime_error { "CoreVM runtime quota exceeded." } {}
-    };
-
-    using Value = uint64_t;
-    using Globals = std::vector<Value>;
-
-    using TraceLogger = std::function<void(Instruction instr, size_t ip, size_t sp)>;
-
-    class Stack
-    { // {{{
-      public:
-        explicit Stack(size_t stackSize) { _stack.reserve(stackSize); }
-
-        void push(Value value) { _stack.push_back(value); }
-
-        Value pop()
-        {
-            Value v = _stack.back();
-            _stack.pop_back();
-            return v;
-        }
-
-        void discard(size_t n)
-        {
-            n = std::min(n, _stack.size());
-            _stack.resize(_stack.size() - n);
-        }
-
-        void rotate(size_t n);
-        void rotate(size_t fp, size_t n);
-
-        size_t size() const { return _stack.size(); }
-
-        Value operator[](int relativeIndex) const
-        {
-            if (relativeIndex < 0)
-            {
-                return _stack[_stack.size() + relativeIndex];
-            }
-            else
-            {
-                return _stack[relativeIndex];
-            }
-        }
-
-        Value& operator[](int relativeIndex)
-        {
-            if (relativeIndex < 0)
-            {
-                return _stack[_stack.size() + relativeIndex];
-            }
-            else
-            {
-                return _stack[relativeIndex];
-            }
-        }
-
-        Value operator[](size_t absoluteIndex) const { return _stack[absoluteIndex]; }
-
-        Value& operator[](size_t absoluteIndex) { return _stack[absoluteIndex]; }
-
-      private:
-        std::vector<Value> _stack;
-    };
-
-    // }}}
-
-  public:
-    /// Result type for run operations that may fail with a RuntimeError
-    using RunResult = std::expected<bool, RuntimeError>;
-
-    Runner(
-        const Function* function, void* userdata, Globals* globals, RuntimeConfig config, TraceLogger logger);
-    Runner(const Function* function,
-           void* userdata,
-           Globals* globals,
-           Quota quota,
-           RuntimeConfig config,
-           TraceLogger logger);
-    ~Runner() = default;
-
-    /// Access the runtime configuration
-    const RuntimeConfig& config() const noexcept { return _config; }
-
-    const Function* function() const noexcept { return _function; }
-
-    const Program* program() const noexcept { return _program; }
-
-    void* userdata() const noexcept { return _userdata; }
-
-    /// Run the function. Returns true if exit code was non-zero.
-    /// On runtime error, prints error and returns true (non-zero exit).
-    bool run();
-
-    /// Run the function with full error reporting.
-    /// Returns expected with bool (true if exit non-zero) or RuntimeError on failure.
-    RunResult runWithResult();
-
-    void suspend();
-    bool resume();
-    void rewind();
-
-    /** Retrieves the last saved program execution offset. */
-    size_t getInstructionPointer() const noexcept { return _ip; }
-
-    /** Retrieves number of elements on stack. */
-    size_t getStackPointer() const noexcept { return _stack.size(); }
-
-    const util::RegExpContext* regexpContext() const noexcept { return &_regexpContext; }
-
-    CoreString* newString(std::string value);
-
-    /// Provides read-only access to the runtime stack (e.g. to inspect alloca values after execution).
-    const Stack& stack() const noexcept { return _stack; }
-
-    /// Allocates a new typed object with the given type ID.
-    TypedObject* allocObject(uint16_t typeId);
-
-    /// Creates a Nil (empty) list with the given element type tag.
-    /// @param elemType The LiteralType of list elements (default Void = unknown).
-    TypedObject* makeNilList(LiteralType elemType = LiteralType::Void);
-
-    /// Creates a Cons cell (head :: tail) with the given element type tag.
-    TypedObject* makeConsCell(uint64_t head, TypedObject* tail, LiteralType elemType = LiteralType::Void);
-
-    /// Creates a Some(value) option with the given inner type tag.
-    TypedObject* makeSomeOption(uint64_t value, LiteralType innerType = LiteralType::Void);
-
-    /// Creates a None option.
-    TypedObject* makeNoneOption();
-
-    /// Creates an Ok(value) result with the given inner type tag.
-    TypedObject* makeOkResult(uint64_t value, LiteralType innerType = LiteralType::Void);
-
-    /// Creates an Error(value) result with the given inner type tag.
-    TypedObject* makeErrorResult(uint64_t value, LiteralType innerType = LiteralType::Void);
-
-    /// Checks if a raw value is a pointer to a known TypedObject allocated by this Runner.
-    [[nodiscard]] bool isKnownObject(uint64_t rawValue) const noexcept;
-
-    /// Checks if a raw value is a pointer to a known CoreString managed by this Runner.
-    [[nodiscard]] bool isKnownString(uint64_t rawValue) const noexcept;
-
-  private:
-    //! consumes @p tokens from quota and raises QuotaExceeded if quota is being exceeded.
-    void consume(Opcode op);
-
-    //! retrieves a pointer to a an empty string constant.
-    const CoreString* emptyString() const { return &*_stringGarbage.begin(); }
-
-    CoreString* catString(const CoreString& a, const CoreString& b);
-
-    Value stack(int si) const { return _stack[si]; }
-
-    CoreNumber getNumber(int si) const { return static_cast<CoreNumber>(_stack[si]); }
-
-    const CoreString& getString(int si) const { return *(CoreString*) _stack[si]; }
-
-    const util::IPAddress& getIPAddress(int si) const { return *(util::IPAddress*) _stack[si]; }
-
-    const util::Cidr& getCidr(int si) const { return *(util::Cidr*) _stack[si]; }
-
-    const util::RegExp& getRegExp(int si) const { return *(util::RegExp*) _stack[si]; }
-
-    const CoreString* getStringPtr(int si) const { return (CoreString*) _stack[si]; }
-
-    const util::Cidr* getCidrPtr(int si) const { return (util::Cidr*) _stack[si]; }
-
-    // Object helpers
-    TypedObject* getObject(int si) const { return reinterpret_cast<TypedObject*>(_stack[si]); }
-
-    void freeObject(TypedObject* obj);
-    const TypeRegistry& typeRegistry() const;
-
-    void push(Value value) { _stack.push(value); }
-
-    Value pop() { return _stack.pop(); }
-
-    void discard(size_t n) { _stack.discard(n); }
-
-    void pushString(const CoreString* value) { push((Value) value); }
-
-    void pushObject(TypedObject* obj) { push(reinterpret_cast<Value>(obj)); }
-
-    RunResult loopWithResult();
-
-    /// Creates a partial Callable by merging existing captures with new partial args.
-    /// Used by IUCALL/IUTCALL when fewer arguments are supplied than the function expects.
-    TypedObject* createPartialCallable(TypedObject* callable, std::vector<Value> const& partialArgs);
-
-    /// Create a RuntimeError at the current instruction's source location
-    [[nodiscard]] RuntimeError makeError(std::string message) const;
-
-    Runner(Runner&) = delete;
-    Runner& operator=(Runner&) = delete;
-
-  private:
-    Quota _quota;
-    RuntimeConfig _config;
-    const Function* _function;
-    TraceLogger _traceLogger;
-
-    /**
-     * We especially keep this ref to prevent ensure function has
-     * access to the program until the end, which is not guaranteed
-     * as it is only having a weak reference to the program (to avoid cycling
-     * references).
-     */
-    const Program* _program;
-
-    //! pointer to the currently evaluated HttpRequest/HttpResponse our case
-    void* _userdata;
-
-    util::RegExpContext _regexpContext;
-
-    State _state; //!< current VM state
-    size_t _ip;   //!< last saved program execution offset
-
-    Stack _stack; //!< runtime stack
-
-    Globals& _globals; //!< runtime global scope
-
-    std::list<std::string> _stringGarbage;
-
-    // Object pool for heap-allocated composite types (Option, Result, etc.)
-    std::vector<std::unique_ptr<uint8_t[]>> _objectPool;
-    size_t _objectAllocCount = 0; //!< Allocation counter for GC threshold
-
-    // Frame pointer for user-defined function calls (UCALL/URET/UTCALL).
-    // For the main function, _fp is always 0, so LOAD/STORE remain backward-compatible.
-    size_t _fp = 0;
-
-    /// Saved call frame for returning from UCALL via URET.
-    struct CallFrame
-    {
-        size_t ip;                      //!< Saved instruction pointer (resume position in caller)
-        const Function* function;       //!< Caller's function
-        size_t fp;                      //!< Caller's frame pointer
-        size_t argsBase;                //!< Stack position where callee's args start (for cleanup on return)
-        TypedObject* lazyObj = nullptr; //!< Non-null when forcing a lazy value (for caching result in URET)
-    };
-
-    /// Call stack for UCALL/URET (user-defined function call frames).
-    std::vector<CallFrame> _callStack;
-
-    /// Maximum call depth for user-defined function calls (prevents stack overflow).
-    static constexpr size_t MaxCallDepth = 10000;
-};
-
-struct MatchCaseDef
-{
-    //!< offset into the string pool (or regexp pool) of the associated program.
-    uint64_t label;
-    //!< program offset into the associated function
-    uint64_t pc;
-
-    MatchCaseDef() = default;
-
-    explicit MatchCaseDef(uint64_t l): label(l), pc(0) {}
-
-    MatchCaseDef(uint64_t l, uint64_t p): label(l), pc(p) {}
-};
-
-class MatchDef
-{
-  public:
-    size_t functionId;
-    MatchClass op; // == =^ =$ =~
-    uint64_t elsePC;
-    std::vector<MatchCaseDef> cases;
-};
-
-class Match
-{
-  public:
-    explicit Match(MatchDef def);
-    virtual ~Match() = default;
-
-    const MatchDef& def() const { return _def; }
-
-    /**
-     * Matches input condition.
-     * \return a code pointer to continue processing
-     */
-    virtual uint64_t evaluate(const CoreString* condition, Runner* env) const = 0;
-
-  protected:
-    MatchDef _def;
-};
-
-/**
- * Provides a pool of constant that can be built dynamically during code
- * generation and accessed effeciently at runtime.
- *
- * @see Program
- */
-class ConstantPool
-{
-  public:
-    using Code = std::vector<Instruction>;
-
-    ConstantPool(const ConstantPool& v) = delete;
-    ConstantPool& operator=(const ConstantPool& v) = delete;
-
-    ConstantPool() = default;
-    ConstantPool(ConstantPool&& from) noexcept = default;
-    ConstantPool& operator=(ConstantPool&& v) noexcept = default;
-
-    // builder
-    size_t makeInteger(CoreNumber value);
-    size_t makeFloat(double value);
-    size_t makeString(const std::string& value);
-    size_t makeIPAddress(const util::IPAddress& value);
-    size_t makeCidr(const util::Cidr& value);
-    size_t makeRegExp(const util::RegExp& value);
-
-    size_t makeIntegerArray(const std::vector<CoreNumber>& elements);
-    size_t makeStringArray(const std::vector<std::string>& elements);
-    size_t makeIPaddrArray(const std::vector<util::IPAddress>& elements);
-    size_t makeCidrArray(const std::vector<util::Cidr>& elements);
-
-    size_t makeMatchDef();
-
-    MatchDef& getMatchDef(size_t id) { return _matchDefs[id]; }
-
-    size_t makeNativeFunction(const std::string& sig);
-    size_t makeNativeFunction(const IRBuiltinFunction* function);
-
-    size_t makeFunction(const std::string& functionName);
-    size_t makeFunction(const IRFunction* function);
-
-    void setModules(const std::vector<std::pair<std::string, std::string>>& modules) { _modules = modules; }
-
-    // accessor
-    [[nodiscard]] CoreNumber getInteger(size_t id) const { return _numbers[id]; }
-
-    [[nodiscard]] double getFloat(size_t id) const { return _floats[id]; }
-
-    [[nodiscard]] const CoreString& getString(size_t id) const { return _strings[id]; }
-
-    [[nodiscard]] const util::IPAddress& getIPAddress(size_t id) const { return _ipaddrs[id]; }
-
-    [[nodiscard]] const util::Cidr& getCidr(size_t id) const { return _cidrs[id]; }
-
-    [[nodiscard]] const util::RegExp& getRegExp(size_t id) const { return _regularExpressions[id]; }
-
-    [[nodiscard]] const std::vector<CoreNumber>& getIntArray(size_t id) const { return _intArrays[id]; }
-
-    [[nodiscard]] const std::vector<std::string>& getStringArray(size_t id) const
-    {
-        return _stringArrays[id];
-    }
-
-    [[nodiscard]] const std::vector<util::IPAddress>& getIPAddressArray(size_t id) const
-    {
-        return _ipaddrArrays[id];
-    }
-
-    [[nodiscard]] const std::vector<util::Cidr>& getCidrArray(size_t id) const { return _cidrArrays[id]; }
-
-    [[nodiscard]] const MatchDef& getMatchDef(size_t id) const { return _matchDefs[id]; }
-
-    [[nodiscard]] const std::pair<std::string, Code>& getFunction(size_t id) const { return _functions[id]; }
-
-    std::pair<std::string, Code>& getFunction(size_t id) { return _functions[id]; }
-
-    size_t setFunction(const std::string& name, Code&& code)
-    {
-        auto id = makeFunction(name);
-        _functions[id].second = std::move(code);
-        return id;
-    }
-
-    // bulk accessors
-    [[nodiscard]] const std::vector<std::pair<std::string, std::string>>& getModules() const
-    {
-        return _modules;
-    }
-
-    [[nodiscard]] const std::vector<std::pair<std::string, Code>>& getFunctions() const { return _functions; }
-
-    /// Location table for each function (parallel to _functions).
-    /// Each entry is a sparse list of (instructionOffset, SourceLocation) pairs.
-    using LocationTable = std::vector<std::pair<size_t, SourceLocation>>;
-
-    void setFunctionLocationTable(size_t functionId, LocationTable table)
-    {
-        if (_functionLocationTables.size() <= functionId)
-            _functionLocationTables.resize(functionId + 1);
-        _functionLocationTables[functionId] = std::move(table);
-    }
-
-    [[nodiscard]] LocationTable const& getFunctionLocationTable(size_t functionId) const
-    {
-        static LocationTable const empty;
-        return functionId < _functionLocationTables.size() ? _functionLocationTables[functionId] : empty;
-    }
-
-    /// Stores the parameter count for a compiled function.
-    void setFunctionParameterCount(size_t functionId, size_t count)
-    {
-        if (_functionParamCounts.size() <= functionId)
-            _functionParamCounts.resize(functionId + 1, 0);
-        _functionParamCounts[functionId] = count;
-    }
-
-    /// Retrieves the parameter count for a compiled function.
-    [[nodiscard]] size_t getFunctionParameterCount(size_t functionId) const
-    {
-        return functionId < _functionParamCounts.size() ? _functionParamCounts[functionId] : 0;
-    }
-
-    [[nodiscard]] const std::vector<MatchDef>& getMatchDefs() const { return _matchDefs; }
-
-    [[nodiscard]] const std::vector<std::string>& getNativeFunctionSignatures() const
-    {
-        return _nativeFunctionSignatures;
-    }
-
-    void dump() const;
-    [[nodiscard]] std::string dumpToString() const;
-
-    // type registry
-    [[nodiscard]] TypeRegistry& typeRegistry() noexcept { return _typeRegistry; }
-
-    [[nodiscard]] const TypeRegistry& typeRegistry() const noexcept { return _typeRegistry; }
-
-  private:
-    // constant primitives
-    std::vector<CoreNumber> _numbers;
-    std::vector<double> _floats;
-    std::vector<std::string> _strings;
-    std::vector<util::IPAddress> _ipaddrs;
-    std::vector<util::Cidr> _cidrs;
-    std::vector<util::RegExp> _regularExpressions;
-
-    // constant arrays
-    std::vector<std::vector<CoreNumber>> _intArrays;
-    std::vector<std::vector<std::string>> _stringArrays;
-    std::vector<std::vector<util::IPAddress>> _ipaddrArrays;
-    std::vector<std::vector<util::Cidr>> _cidrArrays;
-
-    // code data
-    std::vector<std::pair<std::string, std::string>> _modules;
-    std::vector<std::pair<std::string, Code>> _functions;
-    std::vector<LocationTable> _functionLocationTables;
-    std::vector<size_t> _functionParamCounts;
-    std::vector<MatchDef> _matchDefs;
-    std::vector<std::string> _nativeFunctionSignatures;
-
-    // type registry for composite types (Option, Result, etc.)
-    TypeRegistry _typeRegistry;
-};
-
-class Program
-{
-  public:
-    explicit Program(ConstantPool&& cp);
-    Program(Program&) = delete;
-    Program& operator=(Program&) = delete;
-    ~Program() = default;
-
-    const ConstantPool& constants() const noexcept { return _cp; }
-
-    ConstantPool& constants() noexcept { return _cp; }
-
-    /// Provides mutable access to the type registry for runtime type registration.
-    TypeRegistry& mutableTypeRegistry() noexcept { return _cp.typeRegistry(); }
-
-    // accessors to linked data
-    const Match* match(size_t index) const { return _matches[index].get(); }
-
-    Function* function(size_t index) const;
-
-    NativeCallback* nativeFunction(size_t index) const { return _nativeFunctions[index]; }
-
-    // bulk accessors
-    auto matches() { return util::unbox(_matches); }
-
-    std::vector<std::string> functionNames() const;
-    int indexOf(const Function* function) const noexcept;
-    Function* findFunction(const std::string& name) const noexcept;
-
-    bool link(Runtime* runtime, diagnostics::Report* report);
-
-    void dump();
-    [[nodiscard]] std::string dumpToString() const;
-
-    /// Creates a function with the given name and bytecode.
-    /// Useful for dynamically adding pre-compiled functions.
-    using Code = ConstantPool::Code;
-    Function* createFunction(const std::string& name, const Code& instructions);
-
-  private:
-    void setup();
-
-    // builders
-    Function* createFunction(const std::string& name);
-
-  private:
-    ConstantPool _cp;
-
-    // linked data
-    Runtime* _runtime;
-    mutable std::vector<std::unique_ptr<Function>> _functions;
-    std::vector<std::unique_ptr<Match>> _matches;
-    std::vector<NativeCallback*> _nativeFunctions;
-};
-
-class Function
-{
-  public:
-    Function(Program* program, std::string name, std::vector<Instruction> instructions);
-    Function() = default;
-    Function(const Function&) = delete;
-    Function(Function&&) noexcept = default;
-    Function& operator=(const Function&) = delete;
-    Function& operator=(Function&&) noexcept = default;
-    ~Function() = default;
-
-    [[nodiscard]] Program* program() const noexcept { return _program; }
-
-    [[nodiscard]] const std::string& name() const noexcept { return _name; }
-
-    void setName(const std::string& name) { _name = name; }
-
-    [[nodiscard]] size_t stackSize() const noexcept { return _stackSize; }
-
-    [[nodiscard]] const std::vector<Instruction>& code() const noexcept { return _code; }
-
-    void setCode(std::vector<Instruction> code);
-
-#if defined(COREVM_DIRECT_THREADED_VM)
-    [[nodiscard]] const std::vector<uint64_t>& directThreadedCode() const noexcept
-    {
-        return _directThreadedCode;
-    }
-
-    [[nodiscard]] std::vector<uint64_t>& directThreadedCode() noexcept { return _directThreadedCode; }
-#endif
-
-    /// Number of parameters expected by this function (for partial application detection).
-    [[nodiscard]] size_t parameterCount() const noexcept { return _parameterCount; }
-
-    /// Sets the number of parameters expected by this function.
-    void setParameterCount(size_t count) { _parameterCount = count; }
-
-    void disassemble() const noexcept;
-
-    /// Set the sparse location table (only stores locations when they change)
-    /// Each entry is (instructionOffset, SourceLocation)
-    void setLocationTable(std::vector<std::pair<size_t, SourceLocation>> table);
-
-    /// Look up source location for an instruction offset.
-    /// Uses binary search to find the nearest preceding location entry.
-    [[nodiscard]] SourceLocation const& locationOf(size_t offset) const;
-
-  private:
-    Program* _program {};
-    std::string _name;
-    size_t _stackSize {};
-    size_t _parameterCount {};
-    std::vector<Instruction> _code;
-    std::unique_ptr<std::vector<std::pair<size_t, SourceLocation>>> _locationTable;
-#if defined(COREVM_DIRECT_THREADED_VM)
-    std::vector<uint64_t> _directThreadedCode;
-#endif
-};
-
-class Params
-{
-  public:
-    using Value = Runner::Value;
-
-    Params(Runner* caller, int argc): _caller(caller), _argc(argc), _argv(argc + 1) {}
-
-    void setArg(int argi, Value value) { _argv[argi] = value; }
-
-    [[nodiscard]] Runner* caller() const { return _caller; }
-
-    void setResult(bool value) { _argv[0] = value; }
-
-    void setResult(CoreNumber value) { _argv[0] = (Value) value; }
-
-    void setResult(const Function* fn) { _argv[0] = _caller->program()->indexOf(fn); }
-
-    void setResult(const char* str) { _argv[0] = (Value) _caller->newString(str); }
-
-    void setResult(std::string str) { _argv[0] = (Value) _caller->newString(std::move(str)); }
-
-    void setResult(const CoreString* str) { _argv[0] = (Value) str; }
-
-    void setResult(const util::IPAddress* ip) { _argv[0] = (Value) ip; }
-
-    void setResult(const util::Cidr* cidr) { _argv[0] = (Value) cidr; }
-
-    [[deprecated("Use count()")]] [[nodiscard]] int size() const { return _argc; }
-
-    [[nodiscard]] int count() const { return _argc; }
-
-    [[nodiscard]] Value at(size_t i) const { return _argv[i]; }
-
-    [[nodiscard]] Value operator[](size_t i) const { return _argv[i]; }
-
-    [[nodiscard]] Value& operator[](size_t i) { return _argv[i]; }
-
-    [[nodiscard]] bool getBool(size_t offset) const { return at(offset); }
-
-    [[nodiscard]] CoreNumber getInt(size_t offset) const { return at(offset); }
-
-    [[nodiscard]] const CoreString& getString(size_t offset) const { return *(CoreString*) at(offset); }
-
-    [[nodiscard]] Function* getFunction(size_t offset) const
-    {
-        return _caller->program()->function(static_cast<size_t>(at(offset)));
-    }
-
-    [[nodiscard]] const util::IPAddress& getIPAddress(size_t offset) const
-    {
-        return *(util::IPAddress*) at(offset);
-    }
-
-    [[nodiscard]] const util::Cidr& getCidr(size_t offset) const { return *(util::Cidr*) at(offset); }
-
-    /// @brief Retrieves a TypedObject pointer from the argument at the given offset.
-    [[nodiscard]] TypedObject* getObject(size_t offset) const
-    {
-        return reinterpret_cast<TypedObject*>(static_cast<uintptr_t>(at(offset)));
-    }
-
-    [[nodiscard]] const CoreIntArray& getIntArray(size_t offset) const { return *(CoreIntArray*) at(offset); }
-
-    [[nodiscard]] const CoreStringArray& getStringArray(size_t offset) const
-    {
-        return *(CoreStringArray*) at(offset);
-    }
-
-    [[nodiscard]] const CoreIPAddrArray& getIPAddressArray(size_t offset) const
-    {
-        return *(CoreIPAddrArray*) at(offset);
-    }
-
-    [[nodiscard]] const CoreCidrArray& getCidrArray(size_t offset) const
-    {
-        return *(CoreCidrArray*) at(offset);
-    }
-
-    class iterator // NOLINT
-    {              // {{{
-      private:
-        Params* _params;
-        size_t _current;
-
-      public:
-        iterator(Params* p, size_t init): _params(p), _current(init) {}
-
-        iterator(const iterator& v) = default;
-
-        [[nodiscard]] size_t offset() const { return _current; }
-
-        [[nodiscard]] Value get() const { return _params->at(_current); }
-
-        [[nodiscard]] Value& operator*() { return _params->_argv[_current]; }
-
-        [[nodiscard]] const Value& operator*() const { return _params->_argv[_current]; }
-
-        iterator& operator++()
-        {
-            if (_current != static_cast<decltype(_current)>(_params->_argc))
-            {
-                ++_current;
-            }
-            return *this;
-        }
-
-        [[nodiscard]] bool operator==(const iterator& other) const { return _current == other._current; }
-
-        [[nodiscard]] bool operator!=(const iterator& other) const { return _current != other._current; }
-    }; // }}}
-
-    [[nodiscard]] iterator begin() { return iterator(this, std::min(1, _argc)); }
-
-    [[nodiscard]] iterator end() { return iterator(this, _argc); }
-
-  private:
-    Runner* _caller;
-    int _argc;
-    std::vector<Value> _argv;
-};
-
-enum class Attribute : unsigned
-{
-    Experimental = 0x0001,   // implementation is experimental, hence, parser can warn on use
-    NoReturn = 0x0002,       // implementation never returns to program code
-    SideEffectFree = 0x0004, // implementation is side effect free
-};
-
-class Signature
-{
-  private:
-    std::string _name;
-    LiteralType _returnType;
-    std::vector<LiteralType> _args;
-
-  public:
-    Signature();
-    explicit Signature(const std::string& signature);
-    Signature(Signature&&) = default;
-    Signature(const Signature&) = default;
-    Signature& operator=(Signature&&) = default;
-    Signature& operator=(const Signature&) = default;
-
-    void setName(std::string name) { _name = std::move(name); }
-
-    void setReturnType(LiteralType rt) { _returnType = rt; }
-
-    void setArgs(std::vector<LiteralType> args) { _args = std::move(args); }
-
-    [[nodiscard]] const std::string& name() const { return _name; }
-
-    [[nodiscard]] LiteralType returnType() const { return _returnType; }
-
-    [[nodiscard]] const std::vector<LiteralType>& args() const { return _args; }
-
-    [[nodiscard]] std::vector<LiteralType>& args() { return _args; }
-
-    [[nodiscard]] std::string to_s() const;
-
-    [[nodiscard]] bool operator==(const Signature& v) const { return to_s() == v.to_s(); }
-
-    [[nodiscard]] bool operator!=(const Signature& v) const { return to_s() != v.to_s(); }
-
-    [[nodiscard]] bool operator<(const Signature& v) const { return to_s() < v.to_s(); }
-
-    [[nodiscard]] bool operator>(const Signature& v) const { return to_s() > v.to_s(); }
-
-    [[nodiscard]] bool operator<=(const Signature& v) const { return to_s() <= v.to_s(); }
-
-    [[nodiscard]] bool operator>=(const Signature& v) const { return to_s() >= v.to_s(); }
-};
-
-LiteralType typeSignature(char ch);
-char signatureType(LiteralType t);
-
-class NativeCallback
-{
-  public:
-    using Functor = std::function<void(Params& args)>;
-    using Verifier = std::function<bool(Instr*, IRBuilder*)>;
-    using DefaultValue =
-        std::variant<std::monostate, bool, CoreString, CoreNumber, util::IPAddress, util::Cidr, util::RegExp>;
-
-  private:
-    Runtime* _runtime;
-    Verifier _verifier;
-    Functor _function;
-    Signature _signature;
-
-    // function attributes
-    unsigned _attributes;
-
-    // following attribs are irrelevant to the VM but useful for the frontend
-    std::vector<std::string> _names;
-    std::vector<DefaultValue> _defaults;
-
-  public:
-    NativeCallback(Runtime* runtime, std::string name, LiteralType returnType = LiteralType::Void);
-    ~NativeCallback() = default;
-
-    [[nodiscard]] std::string const& name() const noexcept;
-    [[nodiscard]] Signature const& signature() const noexcept;
-
-    // signature builder
-
-    /** Declare the return type. */
-    NativeCallback& returnType(LiteralType type);
-
-    /** Declare a single named parameter by LiteralType (data-driven alternative to param<T>). */
-    NativeCallback& param(LiteralType type, const std::string& name);
-
-    /** Declare a single named parameter without default value. */
-    template <typename T>
-    NativeCallback& param(const std::string& name);
-
-    /** Declare a single named parameter with default value. */
-    template <typename T>
-    NativeCallback& param(const std::string& name, T defaultValue);
-
-    /** Declare ordered parameter signature. */
-    template <typename... Args>
-    NativeCallback& params(Args... args);
-
-    // semantic verifier
-    NativeCallback& verifier(const Verifier& vf);
-    template <typename Class>
-    NativeCallback& verifier(bool (Class::*method)(Instr*, IRBuilder*), Class* obj);
-    template <typename Class>
-    NativeCallback& verifier(bool (Class::*method)(Instr*, IRBuilder*));
-    bool verify(Instr* call, IRBuilder* irBuilder);
-
-    // bind callback
-    NativeCallback& bind(const Functor& cb);
-    template <typename Class>
-    NativeCallback& bind(void (Class::*method)(Params&), Class* obj);
-    template <typename Class>
-    NativeCallback& bind(void (Class::*method)(Params&));
-
-    // named parameter handling
-    [[nodiscard]] bool parametersNamed() const;
-    [[nodiscard]] const std::string& getParamNameAt(size_t i) const;
-    [[nodiscard]] const DefaultValue& getDefaultParamAt(size_t i) const;
-    [[nodiscard]] int findParamByName(const std::string& name) const;
-
-    // attributes
-    NativeCallback& setNoReturn() noexcept;
-    NativeCallback& setReadOnly() noexcept;
-    NativeCallback& setExperimental() noexcept;
-
-    [[nodiscard]] bool getAttribute(Attribute t) const noexcept { return _attributes & unsigned(t); }
-
-    [[nodiscard]] bool isNeverReturning() const noexcept { return getAttribute(Attribute::NoReturn); }
-
-    [[nodiscard]] bool isReadOnly() const noexcept { return getAttribute(Attribute::SideEffectFree); }
-
-    [[nodiscard]] bool isExperimental() const noexcept { return getAttribute(Attribute::Experimental); }
-
-    // runtime
-    void invoke(Params& args) const;
-};
-
-/// A builtin property with getter and/or setter callbacks.
-///
-/// Properties are syntactic sugar over native callbacks. When registered via
-/// Runtime::registerProperty(), both a getter callback (name()T) and a setter
-/// callback (name(T)V) are created automatically so the existing CALL bytecode
-/// works without VM changes.
-class NativeProperty
-{
-  public:
-    using Getter = std::function<void(Params& args)>;
-    using Setter = std::function<void(Params& args)>;
-
-    /// Constructs a named property with the given value type.
-    NativeProperty(std::string name, LiteralType type);
-
-    /// @return The property name.
-    [[nodiscard]] std::string const& name() const noexcept { return _name; }
-
-    /// @return The value type of this property.
-    [[nodiscard]] LiteralType type() const noexcept { return _type; }
-
-    /// @return The human-readable description of this property.
-    [[nodiscard]] std::string const& description() const noexcept { return _description; }
-
-    /// Sets the human-readable description. Returns *this for chaining.
-    NativeProperty& description(std::string desc);
-
-    /// @return true if a getter callback is set.
-    [[nodiscard]] bool hasGetter() const noexcept { return _getter != nullptr; }
-
-    /// @return true if a setter callback is set.
-    [[nodiscard]] bool hasSetter() const noexcept { return _setter != nullptr; }
-
-    /// Sets the getter callback. Returns *this for chaining.
-    NativeProperty& onGet(Getter cb);
-
-    /// Sets the setter callback. Returns *this for chaining.
-    NativeProperty& onSet(Setter cb);
-
-    /// Sets the getter callback using a member function pointer. Returns *this for chaining.
-    template <typename Class>
-    NativeProperty& onGet(void (Class::*method)(Params&), Class* obj)
-    {
-        return onGet([obj, method](Params& args) { (obj->*method)(args); });
-    }
-
-    /// Sets the setter callback using a member function pointer. Returns *this for chaining.
-    template <typename Class>
-    NativeProperty& onSet(void (Class::*method)(Params&), Class* obj)
-    {
-        return onSet([obj, method](Params& args) { (obj->*method)(args); });
-    }
-
-    /// Invokes the getter callback.
-    void invokeGet(Params& args) const;
-
-    /// Invokes the setter callback.
-    void invokeSet(Params& args) const;
-
-  private:
-    std::string _name;
-    LiteralType _type;
-    std::string _description;
-    Getter _getter;
-    Setter _setter;
-};
-
-/// Runtime error with source location for user-friendly error reporting.
-///
-/// Used when the VM encounters an error during execution (e.g., type mismatch,
-/// null dereference). The source location helps users identify the problematic code.
-struct RuntimeError
-{
-    std::string message;
-    SourceLocation location;
-
-    /// Format the error for display to the user
-    [[nodiscard]] std::string format() const
-    {
-        if (location.filename.empty())
-            return std::format("runtime error: {}", message);
-        return std::format("{}:{}:{}: runtime error: {}",
-                           location.filename,
-                           location.begin.line,
-                           location.begin.column,
-                           message);
-    }
-};
-
-//!@}
-
-class PassManager
-{
-  public:
-    using FunctionPass = std::function<bool(IRFunction* function)>;
-
-    PassManager() = default;
-    ~PassManager() = default;
-
-    /** registers given pass to the pass manager.
-     *
-     * @param name uniquely identifiable name of the function pass
-     * @param functionPass callback to invoke to handle the transformation pass
-     *
-     * The callback must return @c true if it modified its input, @c false otherwise.
-     */
-    void registerPass(std::string name, FunctionPass functionPass);
-
-    /** runs passes on a complete program.
-     */
-    void run(IRProgram* program);
-
-    /** runs passes on given function.
-     */
-    void run(IRFunction* function);
-
-    template <typename... Args>
-    void logDebug(std::format_string<Args...> msg, Args... args)
-    {
-        logDebug(std::vformat(msg.get(), std::make_format_args(args...)));
-    }
-
-    void logDebug(const std::string& msg);
-
-  private:
-    std::list<std::pair<std::string, FunctionPass>> _functionPasses;
-};
-
-/**
- * Defines an immutable IR value.
- */
-class Value
-{
-  protected:
-    Value(const Value& v);
-
-  public:
-    Value(LiteralType ty, std::string name);
-    Value(Value&&) noexcept = delete;
-    Value& operator=(Value&&) noexcept = delete;
-    Value& operator=(Value const&) = delete;
-    virtual ~Value();
-
-    [[nodiscard]] LiteralType type() const { return _type; }
-
-    void setType(LiteralType ty) { _type = ty; }
-
-    [[nodiscard]] const std::string& name() const { return _name; }
-
-    void setName(const std::string& n) { _name = n; }
-
-    /**
-     * adds @p user to the list of instructions that are "using" this value.
-     */
-    void addUse(Instr* user);
-
-    /**
-     * removes @p user from the list of instructions that determines the list of
-     * instructions that are using this value.
-     */
-    void removeUse(Instr* user);
-
-    /**
-     * Determines whether or not this value is being used by at least one other
-     * instruction.
-     */
-    [[nodiscard]] bool isUsed() const { return !_uses.empty(); }
-
-    /**
-     * Retrieves a range instructions that are *using* this value.
-     */
-    [[nodiscard]] const std::vector<Instr*>& uses() const { return _uses; }
-
-    [[nodiscard]] size_t useCount() const { return _uses.size(); }
-
-    /**
-     * Replaces all uses of \c this value as operand with value \p newUse instead.
-     *
-     * @param newUse the new value to be used.
-     */
-    void replaceAllUsesWith(Value* newUse);
-
-    [[nodiscard]] virtual std::string to_string() const;
-
-  private:
-    LiteralType _type;
-    std::string _name;
-
-    std::vector<Instr*> _uses; //! list of instructions that <b>use</b> this value.
-};
-
-class Constant: public Value
-{
-  public:
-    Constant(LiteralType ty, std::string name): Value(ty, std::move(name)) {}
-
-    [[nodiscard]] std::string to_string() const override;
-};
-
-class ConstantArray: public Constant
-{
-  public:
-    ConstantArray(LiteralType elementType, std::vector<Constant*> elements, std::string name = ""):
-        Constant(makeArrayType(elementType), std::move(name)), _elements(std::move(elements))
-    {
-    }
-
-    ConstantArray(const std::vector<Constant*>& elements, const std::string& name = ""):
-        Constant(makeArrayType(elements.front()->type()), name), _elements(elements)
-    {
-    }
-
-    ConstantArray(ConstantArray&&) = delete;
-    ConstantArray& operator=(ConstantArray&&) = delete;
-
-    ConstantArray(const ConstantArray&) = delete;
-    ConstantArray& operator=(const ConstantArray&) = delete;
-
-    [[nodiscard]] const std::vector<Constant*>& get() const { return _elements; }
-
-    [[nodiscard]] LiteralType elementType() const { return _elements[0]->type(); }
-
-  private:
-    std::vector<Constant*> _elements;
-
-    LiteralType makeArrayType(LiteralType elementType);
-};
-
-class IRBuiltinFunction: public Constant
-{
-  public:
-    explicit IRBuiltinFunction(const NativeCallback& cb):
-        Constant(cb.signature().returnType(), cb.signature().name()), _native(cb)
-    {
-    }
-
-    [[nodiscard]] const Signature& signature() const { return _native.signature(); }
-
-    [[nodiscard]] const NativeCallback& getNative() const { return _native; }
-
-  private:
-    const NativeCallback& _native;
-};
-
-template <typename T, const LiteralType Ty>
-class ConstantValue: public Constant
-{
-  public:
-    ConstantValue(T value, const std::string& name = ""): Constant(Ty, name), _value(std::move(value)) {}
-
-    [[nodiscard]] T get() const { return _value; }
-
-    [[nodiscard]] std::string to_string() const override
-    {
-        auto valueStr = [&]() -> std::string {
-            if constexpr (std::is_same_v<T, std::string>)
-                return _value;
-            else if constexpr (std::is_same_v<T, bool>)
-                return _value ? "true" : "false";
-            else if constexpr (std::is_arithmetic_v<T>)
-                return std::to_string(_value);
-            else if constexpr (requires { _value.str(); })
-                return _value.str();
-            else if constexpr (requires { _value.pattern(); })
-                return _value.pattern();
-            else
-                return "?";
-        }();
-        return "Constant '" + std::string(name()) + "': " + tos(type()) + " = " + valueStr;
-    }
-
-  private:
-    T _value;
-};
-
-using ConstantInt = ConstantValue<int64_t, LiteralType::Number>;
-using ConstantBoolean = ConstantValue<bool, LiteralType::Boolean>;
-using ConstantFloat = ConstantValue<double, LiteralType::Float>;
-using ConstantString = ConstantValue<std::string, LiteralType::String>;
-using ConstantIP = ConstantValue<util::IPAddress, LiteralType::IPAddress>;
-using ConstantCidr = ConstantValue<util::Cidr, LiteralType::Cidr>;
-using ConstantRegExp = ConstantValue<util::RegExp, LiteralType::RegExp>;
-
-/**
- * Base class for native instructions.
- *
- * An instruction is derived from base class \c Value because its result can be
- * used as an operand for other instructions.
- *
- * @see IRBuilder
- * @see BasicBlock
- * @see IRFunction
- */
-class Instr: public Value
-{
-  protected:
-    Instr(const Instr& v);
-
-  public:
-    Instr(LiteralType ty, const std::vector<Value*>& ops = {}, const std::string& name = "");
-    ~Instr() override;
-
-    /**
-     * Retrieves parent basic block this instruction is part of.
-     */
-    [[nodiscard]] BasicBlock* getBasicBlock() const { return _basicBlock; }
-
-    /**
-     * Read-only access to operands.
-     */
-    [[nodiscard]] const std::vector<Value*>& operands() const { return _operands; }
-
-    /**
-     * Retrieves n'th operand at given \p index.
-     */
-    [[nodiscard]] Value* operand(size_t index) const { return _operands[index]; }
-
-    /**
-     * Adds given operand \p value to the end of the operand list.
-     */
-    void addOperand(Value* value);
-
-    /**
-     * Sets operand at index \p i to given \p value.
-     *
-     * This operation will potentially replace the value that has been at index \p
-     *i before,
-     * properly unlinking it from any uses or successor/predecessor links.
-     */
-    Value* setOperand(size_t i, Value* value);
-
-    /**
-     * Replaces \p old operand with \p replacement.
-     *
-     * @param old value to replace
-     * @param replacement new value to put at the offset of \p old.
-     *
-     * @returns number of actual performed replacements.
-     */
-    size_t replaceOperand(Value* old, Value* replacement);
-
-    /**
-     * Clears out all operands.
-     *
-     * @see addOperand()
-     */
-    void clearOperands();
-
-    /**
-     * Replaces this instruction with the given @p newInstr.
-     *
-     * @returns ownership of this instruction.
-     */
-    std::unique_ptr<Instr> replace(std::unique_ptr<Instr> newInstr);
-
-    /**
-     * Clones given instruction.
-     *
-     * This will not clone any of its operands but reference them.
-     */
-    virtual std::unique_ptr<Instr> clone() = 0;
-
-    /**
-     * Generic extension interface.
-     *
-     * @param v extension to pass this instruction to.
-     *
-     * @see InstructionVisitor
-     */
-    virtual void accept(InstructionVisitor& v) = 0;
-
-    /// Set the source location for this instruction (for error reporting)
-    void setSourceLocation(SourceLocation loc) { _sourceLocation = std::move(loc); }
-
-    /// Get the source location of this instruction
-    [[nodiscard]] SourceLocation const& sourceLocation() const noexcept { return _sourceLocation; }
-
-  protected:
-    void dumpOne(const char* mnemonic);
-    [[nodiscard]] std::string formatOne(std::string mnemonic) const;
-
-    void setParent(BasicBlock* bb) { _basicBlock = bb; }
-
-    friend class BasicBlock;
-
-  private:
-    BasicBlock* _basicBlock;
-    std::vector<Value*> _operands;
-    SourceLocation _sourceLocation;
-};
-
-class NopInstr: public Instr
-{
-  public:
-    NopInstr(): Instr(LiteralType::Void, {}, "nop") {}
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-/**
- * Allocates an array of given type and elements.
- */
-class AllocaInstr: public Instr
-{
-  private:
-    static LiteralType computeType(LiteralType elementType, Value* size)
-    { // {{{
-        if (auto* n = dynamic_cast<ConstantInt*>(size))
-        {
-            if (n->get() == 1)
-                return elementType;
-        }
-
-        switch (elementType)
-        {
-            case LiteralType::Number: return LiteralType::IntArray;
-            case LiteralType::String: return LiteralType::StringArray;
-            default: return LiteralType::Void;
-        }
-    } // }}}
-
-  public:
-    AllocaInstr(LiteralType ty, Value* n, const std::string& name): Instr(ty, { n }, name) {}
-
-    [[nodiscard]] LiteralType elementType() const
-    {
-        switch (type())
-        {
-            case LiteralType::StringArray: return LiteralType::String;
-            case LiteralType::IntArray: return LiteralType::Number;
-            default: return LiteralType::Void;
-        }
-    }
-
-    [[nodiscard]] Value* arraySize() const { return operands()[0]; }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-class StoreInstr: public Instr
-{
-  public:
-    StoreInstr(Value* variable, ConstantInt* index, Value* source, const std::string& name):
-        Instr(LiteralType::Void, { variable, index, source }, name)
-    {
-    }
-
-    [[nodiscard]] Value* variable() const { return operand(0); }
-
-    [[nodiscard]] ConstantInt* index() const { return static_cast<ConstantInt*>(operand(1)); }
-
-    [[nodiscard]] Value* source() const { return operand(2); }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-class RegExpGroupInstr: public Instr
-{
-  public:
-    RegExpGroupInstr(ConstantInt* groupId, const std::string& name):
-        Instr { LiteralType::String, { groupId }, name }
-    {
-    }
-
-    [[nodiscard]] ConstantInt* groupId() const { return static_cast<ConstantInt*>(operand(0)); }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-class LoadInstr: public Instr
-{
-  public:
-    LoadInstr(Value* variable, const std::string& name): Instr(variable->type(), { variable }, name) {}
-
-    [[nodiscard]] Value* variable() const { return operand(0); }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-class CallInstr: public Instr
-{
-  public:
-    CallInstr(const std::vector<Value*>& args, const std::string& name);
-    CallInstr(IRBuiltinFunction* callee, const std::vector<Value*>& args, const std::string& name);
-
-    [[nodiscard]] IRBuiltinFunction* callee() const { return (IRBuiltinFunction*) operand(0); }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-// =============================================================================
-// User-Defined Function Call Instructions (UCALL/URET/UTCALL)
-// =============================================================================
-
-/// Calls a user-defined function compiled as a separate IRFunction.
-/// Operands: [arg0, arg1, ...]. Return type reflects the callee's body result type.
-class FunctionCallInstr: public Instr
-{
-  public:
-    FunctionCallInstr(IRFunction* callee,
-                      std::vector<Value*> args,
-                      const std::string& name,
-                      LiteralType returnType = LiteralType::Void);
-
-    /// The target function to call.
-    [[nodiscard]] IRFunction* callee() const { return _callee; }
-
-    /// Number of arguments (excluding the function operand).
-    [[nodiscard]] size_t argc() const { return operands().size(); }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-
-  private:
-    IRFunction* _callee;
-};
-
-/// Returns from a user-defined function back to the caller.
-/// Not a TerminateInstr because it doesn't branch within the function's CFG--
-/// the VM handles control transfer.
-class FunctionRetInstr: public Instr
-{
-  public:
-    explicit FunctionRetInstr(Value* result, const std::string& name = "");
-
-    /// The value to return to the caller.
-    [[nodiscard]] Value* result() const { return operands().empty() ? nullptr : operand(0); }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-/// Tail-calls a user-defined function, reusing the current call frame.
-/// Operands: [arg0, arg1, ...]. Must be in tail position.
-class TailCallInstr: public Instr
-{
-  public:
-    TailCallInstr(IRFunction* callee, std::vector<Value*> args, const std::string& name);
-
-    /// The target function to tail-call.
-    [[nodiscard]] IRFunction* callee() const { return _callee; }
-
-    /// Number of arguments.
-    [[nodiscard]] size_t argc() const { return operands().size(); }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-
-  private:
-    IRFunction* _callee;
-};
-
-/// Produces the runtime function ID of a given IRFunction as a Number value.
-/// Used to store function references in objects (e.g., lazy thunks).
-class FunctionRefInstr: public Instr
-{
-  public:
-    /// @param function The IRFunction whose runtime ID is needed.
-    /// @param name Optional IR name for the result.
-    FunctionRefInstr(IRFunction* function, const std::string& name):
-        Instr(LiteralType::Number, {}, name), _function(function)
-    {
-    }
-
-    /// The referenced function.
-    [[nodiscard]] IRFunction* function() const { return _function; }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-
-  private:
-    IRFunction* _function;
-};
-
-/// Forces evaluation of a lazy value. Reads the thunk function ID from the
-/// lazy object at runtime and invokes it via the LFORCE opcode.
-class LazyForceInstr: public Instr
-{
-  public:
-    /// @param lazyObj The lazy object value to force.
-    /// @param name Optional IR name for the result.
-    LazyForceInstr(Value* lazyObj, const std::string& name): Instr(LiteralType::Void, { lazyObj }, name) {}
-
-    /// The lazy object operand.
-    [[nodiscard]] Value* lazyObj() const { return operand(0); }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-/// Indirect call through a Callable object. The callable is resolved at runtime
-/// to a function ID + captured variables. Operand 0 = callable, operands 1..N = explicit args.
-class IndirectCallInstr: public Instr
-{
-  public:
-    /// @param callable The Callable object value to invoke.
-    /// @param args The explicit arguments to pass (captures are inside the callable).
-    /// @param name Optional IR name for the result.
-    IndirectCallInstr(Value* callable, std::vector<Value*> args, const std::string& name);
-
-    /// The callable object operand.
-    [[nodiscard]] Value* callable() const { return operand(0); }
-
-    /// Number of explicit arguments (excluding the callable).
-    [[nodiscard]] size_t argc() const { return operands().size() - 1; }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-/// Indirect tail call through a Callable object. Same semantics as IndirectCallInstr
-/// but reuses the current frame (no CallFrame push). Result type is Void since
-/// tail calls transfer control and don't produce a value in the current frame.
-class IndirectTailCallInstr: public Instr
-{
-  public:
-    /// @param callable The Callable object value to invoke.
-    /// @param args The explicit arguments to pass (captures are inside the callable).
-    /// @param name Optional IR name.
-    IndirectTailCallInstr(Value* callable, std::vector<Value*> args, const std::string& name);
-
-    /// The callable object operand.
-    [[nodiscard]] Value* callable() const { return operand(0); }
-
-    /// Number of explicit arguments (excluding the callable).
-    [[nodiscard]] size_t argc() const { return operands().size() - 1; }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-class CastInstr: public Instr
-{
-  public:
-    CastInstr(LiteralType resultType, Value* op, const std::string& name): Instr(resultType, { op }, name) {}
-
-    [[nodiscard]] Value* source() const { return operand(0); }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-template <const UnaryOperator Operator, const LiteralType ResultType>
-class UnaryInstr: public Instr
-{
-  public:
-    UnaryInstr(Value* op, const std::string& name): Instr(ResultType, { op }, name), _operator(Operator) {}
-
-    [[nodiscard]] UnaryOperator op() const { return _operator; }
-
-    [[nodiscard]] std::string to_string() const override { return formatOne(cstr(_operator)); }
-
-    [[nodiscard]] std::unique_ptr<Instr> clone() override
-    {
-        return std::make_unique<UnaryInstr<Operator, ResultType>>(operand(0), name());
-    }
-
-    void accept(InstructionVisitor& v) override { v.visit(*this); }
-
-  private:
-    UnaryOperator _operator;
-};
-
-template <const BinaryOperator Operator, const LiteralType ResultType>
-class BinaryInstr: public Instr
-{
-  public:
-    BinaryInstr(Value* lhs, Value* rhs, const std::string& name):
-        Instr(ResultType, { lhs, rhs }, name), _operator(Operator)
-    {
-    }
-
-    [[nodiscard]] BinaryOperator op() const { return _operator; }
-
-    [[nodiscard]] std::string to_string() const override { return formatOne(cstr(_operator)); }
-
-    [[nodiscard]] std::unique_ptr<Instr> clone() override
-    {
-        return std::make_unique<BinaryInstr<Operator, ResultType>>(operand(0), operand(1), name());
-    }
-
-    void accept(InstructionVisitor& v) override { v.visit(*this); }
-
-  private:
-    BinaryOperator _operator;
-};
-
-/**
- * Creates a PHI (phoney) instruction.
- *
- * Creates a synthetic instruction that purely informs the target register
- *allocator
- * to allocate the very same register for all given operands,
- * which is then used across all their basic blocks.
- */
-class PhiNode: public Instr
-{
-  public:
-    PhiNode(const std::vector<Value*>& ops, const std::string& name);
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-class TerminateInstr: public Instr
-{
-  protected:
-    TerminateInstr(const TerminateInstr& v) = default;
-
-  public:
-    TerminateInstr(const std::vector<Value*>& ops): Instr(LiteralType::Void, ops, "") {}
-};
-
-/**
- * Conditional branch instruction.
- *
- * Creates a terminate instruction that transfers control to one of the two
- * given alternate basic blocks, depending on the given input condition.
- */
-class CondBrInstr: public TerminateInstr
-{
-  public:
-    /**
-     * Initializes the object.
-     *
-     * @param cond input condition that (if true) causes \p trueBlock to be jumped
-     *to, \p falseBlock otherwise.
-     * @param trueBlock basic block to run if input condition evaluated to true.
-     * @param falseBlock basic block to run if input condition evaluated to false.
-     */
-    CondBrInstr(Value* cond, BasicBlock* trueBlock, BasicBlock* falseBlock);
-
-    [[nodiscard]] Value* condition() const { return operands()[0]; }
-
-    [[nodiscard]] BasicBlock* trueBlock() const { return (BasicBlock*) operands()[1]; }
-
-    [[nodiscard]] BasicBlock* falseBlock() const { return (BasicBlock*) operands()[2]; }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-/**
- * Unconditional jump instruction.
- */
-class BrInstr: public TerminateInstr
-{
-  public:
-    explicit BrInstr(BasicBlock* targetBlock);
-
-    [[nodiscard]] BasicBlock* targetBlock() const { return (BasicBlock*) operands()[0]; }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-/**
- * Function-return instruction.
- */
-class RetInstr: public TerminateInstr
-{
-  public:
-    RetInstr(Value* result);
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-/**
- * Match instruction, implementing the CoreVM match-keyword.
- *
- * <li>operand[0] - condition</li>
- * <li>operand[1] - default block</li>
- * <li>operand[2n+2] - case label</li>
- * <li>operand[2n+3] - case block</li>
- */
-class MatchInstr: public TerminateInstr
-{
-  public:
-    MatchInstr(const MatchInstr&);
-    MatchInstr(MatchClass op, Value* cond);
-
-    [[nodiscard]] MatchClass op() const { return _op; }
-
-    [[nodiscard]] Value* condition() const { return operand(0); }
-
-    void addCase(Constant* label, BasicBlock* code);
-    [[nodiscard]] std::vector<std::pair<Constant*, BasicBlock*>> cases() const;
-
-    [[nodiscard]] BasicBlock* elseBlock() const;
-    void setElseBlock(BasicBlock* code);
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-
-  private:
-    MatchClass _op;
-    std::vector<std::pair<Constant*, BasicBlock*>> _cases;
-};
-
-// =============================================================================
-// Object Instructions (for composite types: Option, Result, tuples, closures)
-// =============================================================================
-
-/// Allocates a new typed object on the heap.
-/// @param typeId The type ID from the TypeRegistry
-class ObjAllocInstr: public Instr
-{
-  public:
-    ObjAllocInstr(ConstantInt* typeId, const std::string& name): Instr(LiteralType::Object, { typeId }, name)
-    {
-    }
-
-    [[nodiscard]] ConstantInt* typeId() const { return static_cast<ConstantInt*>(operand(0)); }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-/// Increments the reference count of an object.
-class ObjRetainInstr: public Instr
-{
-  public:
-    ObjRetainInstr(Value* object, const std::string& name): Instr(LiteralType::Object, { object }, name) {}
-
-    [[nodiscard]] Value* object() const { return operand(0); }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-/// Decrements the reference count and frees if zero.
-/// Takes an AllocaInstr* (storage location) directly to avoid cross-block value tracking issues.
-class ObjReleaseInstr: public Instr
-{
-  public:
-    ObjReleaseInstr(AllocaInstr* storage, const std::string& name):
-        Instr(LiteralType::Void, { storage }, name)
-    {
-    }
-
-    [[nodiscard]] AllocaInstr* storage() const { return static_cast<AllocaInstr*>(operand(0)); }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-/// Gets the tag (variant discriminant) from a sum type object.
-class ObjGetTagInstr: public Instr
-{
-  public:
-    ObjGetTagInstr(Value* object, const std::string& name): Instr(LiteralType::Number, { object }, name) {}
-
-    [[nodiscard]] Value* object() const { return operand(0); }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-/// Sets the tag (variant discriminant) on a sum type object.
-class ObjSetTagInstr: public Instr
-{
-  public:
-    ObjSetTagInstr(Value* object, Value* tag, const std::string& name):
-        Instr(LiteralType::Object, { object, tag }, name)
-    {
-    }
-
-    [[nodiscard]] Value* object() const { return operand(0); }
-
-    [[nodiscard]] Value* tag() const { return operand(1); }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-/// Gets a slot value from an object.
-class ObjGetSlotInstr: public Instr
-{
-  public:
-    ObjGetSlotInstr(Value* object, ConstantInt* slotIndex, const std::string& name):
-        Instr(LiteralType::Void, { object, slotIndex }, name) // Void because actual type depends on content
-    {
-    }
-
-    [[nodiscard]] Value* object() const { return operand(0); }
-
-    [[nodiscard]] ConstantInt* slotIndex() const { return static_cast<ConstantInt*>(operand(1)); }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-/// Sets a slot value in an object.
-class ObjSetSlotInstr: public Instr
-{
-  public:
-    ObjSetSlotInstr(Value* object, ConstantInt* slotIndex, Value* value, const std::string& name):
-        Instr(LiteralType::Object, { object, slotIndex, value }, name)
-    {
-    }
-
-    [[nodiscard]] Value* object() const { return operand(0); }
-
-    [[nodiscard]] ConstantInt* slotIndex() const { return static_cast<ConstantInt*>(operand(1)); }
-
-    [[nodiscard]] Value* value() const { return operand(2); }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-/// Gets the type ID of an object.
-class ObjTypeIdInstr: public Instr
-{
-  public:
-    ObjTypeIdInstr(Value* object, const std::string& name): Instr(LiteralType::Number, { object }, name) {}
-
-    [[nodiscard]] Value* object() const { return operand(0); }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-/// Checks if an object is of a specific type.
-class ObjIsTypeInstr: public Instr
-{
-  public:
-    ObjIsTypeInstr(Value* object, ConstantInt* typeId, const std::string& name):
-        Instr(LiteralType::Boolean, { object, typeId }, name)
-    {
-    }
-
-    [[nodiscard]] Value* object() const { return operand(0); }
-
-    [[nodiscard]] ConstantInt* typeId() const { return static_cast<ConstantInt*>(operand(1)); }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-/// Dynamic value comparison instruction.
-/// Compares two values as numbers regardless of their compile-time type.
-/// Used for pattern matching with dynamically-typed values from OGETSLOT.
-class VCmpEQInstr: public Instr
-{
-  public:
-    VCmpEQInstr(Value* lhs, Value* rhs, const std::string& name):
-        Instr(LiteralType::Boolean, { lhs, rhs }, name)
-    {
-    }
-
-    [[nodiscard]] Value* lhs() const { return operand(0); }
-
-    [[nodiscard]] Value* rhs() const { return operand(1); }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-/// Dynamic not-equal comparison instruction.
-class VCmpNEInstr: public Instr
-{
-  public:
-    VCmpNEInstr(Value* lhs, Value* rhs, const std::string& name):
-        Instr(LiteralType::Boolean, { lhs, rhs }, name)
-    {
-    }
-
-    [[nodiscard]] Value* lhs() const { return operand(0); }
-
-    [[nodiscard]] Value* rhs() const { return operand(1); }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-/// Dynamic less-than comparison instruction.
-class VCmpLTInstr: public Instr
-{
-  public:
-    VCmpLTInstr(Value* lhs, Value* rhs, const std::string& name):
-        Instr(LiteralType::Boolean, { lhs, rhs }, name)
-    {
-    }
-
-    [[nodiscard]] Value* lhs() const { return operand(0); }
-
-    [[nodiscard]] Value* rhs() const { return operand(1); }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-/// Dynamic less-than-or-equal comparison instruction.
-class VCmpLEInstr: public Instr
-{
-  public:
-    VCmpLEInstr(Value* lhs, Value* rhs, const std::string& name):
-        Instr(LiteralType::Boolean, { lhs, rhs }, name)
-    {
-    }
-
-    [[nodiscard]] Value* lhs() const { return operand(0); }
-
-    [[nodiscard]] Value* rhs() const { return operand(1); }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-/// Dynamic greater-than comparison instruction.
-class VCmpGTInstr: public Instr
-{
-  public:
-    VCmpGTInstr(Value* lhs, Value* rhs, const std::string& name):
-        Instr(LiteralType::Boolean, { lhs, rhs }, name)
-    {
-    }
-
-    [[nodiscard]] Value* lhs() const { return operand(0); }
-
-    [[nodiscard]] Value* rhs() const { return operand(1); }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-/// Dynamic greater-than-or-equal comparison instruction.
-class VCmpGEInstr: public Instr
-{
-  public:
-    VCmpGEInstr(Value* lhs, Value* rhs, const std::string& name):
-        Instr(LiteralType::Boolean, { lhs, rhs }, name)
-    {
-    }
-
-    [[nodiscard]] Value* lhs() const { return operand(0); }
-
-    [[nodiscard]] Value* rhs() const { return operand(1); }
-
-    [[nodiscard]] std::string to_string() const override;
-    [[nodiscard]] std::unique_ptr<Instr> clone() override;
-    void accept(InstructionVisitor& v) override;
-};
-
-class IsSameInstruction: public InstructionVisitor
-{
-  public:
-    static bool test(Instr* a, Instr* b);
-    static bool isSameOperands(Instr* a, Instr* b);
-
-  private:
-    Instr* _other;
-    bool _result;
-
-  protected:
-    explicit IsSameInstruction(Instr* a);
-
-    void visit(NopInstr& instr) override;
-
-    // storage
-    void visit(AllocaInstr& instr) override;
-    void visit(StoreInstr& instr) override;
-    void visit(LoadInstr& instr) override;
-    void visit(PhiNode& instr) override;
-
-    // calls
-    void visit(CallInstr& instr) override;
-
-    void visit(FunctionCallInstr& instr) override;
-    void visit(FunctionRetInstr& instr) override;
-    void visit(TailCallInstr& instr) override;
-    void visit(IndirectCallInstr& instr) override;
-    void visit(IndirectTailCallInstr& instr) override;
-
-    // terminator
-    void visit(CondBrInstr& instr) override;
-    void visit(BrInstr& instr) override;
-    void visit(RetInstr& instr) override;
-    void visit(MatchInstr& instr) override;
-
-    // regexp
-    void visit(RegExpGroupInstr& instr) override;
-
-    // type cast
-    void visit(CastInstr& instr) override;
-
-    // numeric
-    void visit(INegInstr& instr) override;
-    void visit(INotInstr& instr) override;
-    void visit(IAddInstr& instr) override;
-    void visit(ISubInstr& instr) override;
-    void visit(IMulInstr& instr) override;
-    void visit(IDivInstr& instr) override;
-    void visit(IRemInstr& instr) override;
-    void visit(IPowInstr& instr) override;
-    void visit(IAndInstr& instr) override;
-    void visit(IOrInstr& instr) override;
-    void visit(IXorInstr& instr) override;
-    void visit(IShlInstr& instr) override;
-    void visit(IShrInstr& instr) override;
-    void visit(ICmpEQInstr& instr) override;
-    void visit(ICmpNEInstr& instr) override;
-    void visit(ICmpLEInstr& instr) override;
-    void visit(ICmpGEInstr& instr) override;
-    void visit(ICmpLTInstr& instr) override;
-    void visit(ICmpGTInstr& instr) override;
-
-    // boolean
-    void visit(BNotInstr& instr) override;
-    void visit(BAndInstr& instr) override;
-    void visit(BOrInstr& instr) override;
-    void visit(BXorInstr& instr) override;
-
-    // string
-    void visit(SLenInstr& instr) override;
-    void visit(SIsEmptyInstr& instr) override;
-    void visit(SAddInstr& instr) override;
-    void visit(SSubStrInstr& instr) override;
-    void visit(SCmpEQInstr& instr) override;
-    void visit(SCmpNEInstr& instr) override;
-    void visit(SCmpLEInstr& instr) override;
-    void visit(SCmpGEInstr& instr) override;
-    void visit(SCmpLTInstr& instr) override;
-    void visit(SCmpGTInstr& instr) override;
-    void visit(SCmpREInstr& instr) override;
-    void visit(SCmpBegInstr& instr) override;
-    void visit(SCmpEndInstr& instr) override;
-    void visit(SInInstr& instr) override;
-
-    // ip
-    void visit(PCmpEQInstr& instr) override;
-    void visit(PCmpNEInstr& instr) override;
-    void visit(PInCidrInstr& instr) override;
-
-    // object operations
-    void visit(ObjAllocInstr& instr) override;
-    void visit(ObjRetainInstr& instr) override;
-    void visit(ObjReleaseInstr& instr) override;
-    void visit(ObjGetTagInstr& instr) override;
-    void visit(ObjSetTagInstr& instr) override;
-    void visit(ObjGetSlotInstr& instr) override;
-    void visit(ObjSetSlotInstr& instr) override;
-    void visit(ObjTypeIdInstr& instr) override;
-    void visit(ObjIsTypeInstr& instr) override;
-    void visit(VCmpEQInstr& instr) override;
-    void visit(VCmpNEInstr& instr) override;
-    void visit(VCmpLTInstr& instr) override;
-    void visit(VCmpLEInstr& instr) override;
-    void visit(VCmpGTInstr& instr) override;
-    void visit(VCmpGEInstr& instr) override;
-
-    // float
-    void visit(FNegInstr& instr) override;
-    void visit(FAddInstr& instr) override;
-    void visit(FSubInstr& instr) override;
-    void visit(FMulInstr& instr) override;
-    void visit(FDivInstr& instr) override;
-    void visit(FRemInstr& instr) override;
-    void visit(FPowInstr& instr) override;
-    void visit(FCmpEQInstr& instr) override;
-    void visit(FCmpNEInstr& instr) override;
-    void visit(FCmpLEInstr& instr) override;
-    void visit(FCmpGEInstr& instr) override;
-    void visit(FCmpLTInstr& instr) override;
-    void visit(FCmpGTInstr& instr) override;
-
-    // lazy evaluation
-    void visit(FunctionRefInstr& instr) override;
-    void visit(LazyForceInstr& instr) override;
-};
-
-class IRFunction: public Constant
-{
-  public:
-    IRFunction(const std::string& name, IRProgram* parent);
-    ~IRFunction() override;
-
-    IRFunction(IRFunction&&) = delete;
-    IRFunction& operator=(IRFunction&&) = delete;
-
-    BasicBlock* createBlock(const std::string& name = "");
-
-    [[nodiscard]] IRProgram* getProgram() const { return _program; }
-
-    void setParent(IRProgram* prog) { _program = prog; }
-
-    void dump();
-    [[nodiscard]] std::string dumpToString() const;
-
-    [[nodiscard]] bool empty() const noexcept { return _blocks.empty(); }
-
-    /// Number of parameters expected by this function (for UCALL frame setup).
-    /// Zero for the main function, >0 for compiled F# functions.
-    [[nodiscard]] size_t parameterCount() const noexcept { return _parameterCount; }
-
-    void setParameterCount(size_t count) { _parameterCount = count; }
-
-    auto basicBlocks() { return util::unbox(_blocks); }
-
-    [[nodiscard]] BasicBlock* getEntryBlock() const { return _blocks.front().get(); }
-
-    void setEntryBlock(BasicBlock* bb);
-
-    /**
-     * Unlinks and deletes given basic block \p bb from function.
-     *
-     * @note \p bb will be a dangling pointer after this call.
-     */
-    void erase(BasicBlock* bb);
-
-    bool isAfter(const BasicBlock* bb, const BasicBlock* afterThat) const;
-    void moveAfter(const BasicBlock* moveable, const BasicBlock* afterThat);
-    void moveBefore(const BasicBlock* moveable, const BasicBlock* beforeThat);
-
-    /**
-     * Performs given transformation on this function.
-     */
-    template <typename TheFunctionPass, typename... Args>
-    size_t transform(Args&&... args)
-    {
-        return TheFunctionPass(std::forward(args)...).run(this);
-    }
-
-    /**
-     * Performs sanity checks on internal data structures.
-     *
-     * This call does not return any success or failure as every failure is
-     * considered fatal and will cause the program to exit with diagnostics
-     * as this is most likely caused by an application programming error.
-     *
-     * @note Always call this on completely defined functions and never on
-     * half-contructed ones.
-     */
-    void verify();
-
-  private:
-    IRProgram* _program;
-    std::list<std::unique_ptr<BasicBlock>> _blocks;
-    size_t _parameterCount = 0; ///< Number of parameters (0 for main function)
-
-    friend class IRBuilder;
-};
-
-class IRProgram
-{
-  public:
-    IRProgram();
-    ~IRProgram();
-
-    void dump();
-    [[nodiscard]] std::string dumpToString() const;
-
-    ConstantBoolean* getBoolean(bool literal) { return literal ? &_trueLiteral : &_falseLiteral; }
-
-    ConstantInt* get(int64_t literal) { return get<ConstantInt>(_numbers, literal); }
-
-    ConstantFloat* getFloat(double literal) { return get<ConstantFloat>(_floats, literal); }
-
-    ConstantString* get(const std::string& literal) { return get<ConstantString>(_strings, literal); }
-
-    ConstantIP* get(const util::IPAddress& literal) { return get<ConstantIP>(_ipaddrs, literal); }
-
-    ConstantCidr* get(const util::Cidr& literal) { return get<ConstantCidr>(_cidrs, literal); }
-
-    ConstantRegExp* get(const util::RegExp& literal) { return get<ConstantRegExp>(_regexps, literal); }
-
-    ConstantArray* get(const std::vector<Constant*>& elems)
-    {
-        return get<ConstantArray>(_constantArrays, elems);
-    }
-
-    IRBuiltinFunction* getBuiltinFunction(const NativeCallback& cb)
-    {
-        for (const auto& builtinFunction: _builtinFunctions)
-            if (builtinFunction->signature() == cb.signature())
-                return builtinFunction.get();
-
-        _builtinFunctions.emplace_back(std::make_unique<IRBuiltinFunction>(cb));
-        return _builtinFunctions.back().get();
-    }
-
-    template <typename T, typename U>
-    T* get(std::vector<T>& table, U&& literal);
-    template <typename T, typename U>
-    T* get(std::vector<std::unique_ptr<T>>& table, U&& literal);
-
-    void addImport(const std::string& name, const std::string& path) { _modules.emplace_back(name, path); }
-
-    void setModules(const std::vector<std::pair<std::string, std::string>>& modules) { _modules = modules; }
-
-    [[nodiscard]] const std::vector<std::pair<std::string, std::string>>& modules() const { return _modules; }
-
-    auto functions() { return util::unbox(_functions); }
-
-    IRFunction* findFunction(const std::string& name)
-    {
-        for (IRFunction* fn: functions())
-            if (fn->name() == name)
-                return fn;
-
-        return nullptr;
-    }
-
-    IRFunction* createFunction(const std::string& name);
-
-    /// Removes the given function from the program. The pointer becomes invalid after this call.
-    void removeFunction(IRFunction* function);
-
-    /**
-     * Performs given transformation on all functions by given type.
-     */
-    template <typename TheFunctionPass, typename... Args>
-    size_t transform(Args&&... args)
-    {
-        size_t count = 0;
-        for (IRFunction* fn: functions())
-        {
-            count += fn->transform<TheFunctionPass>(args...);
-        }
-        return count;
-    }
-
-    /// Describes a custom product type to be registered at target code generation time.
-    struct CustomProductType
-    {
-        std::string name;
-        std::vector<FieldInfo> fields;
-        uint16_t assignedId;    ///< Pre-assigned type ID used in OALLOC instructions
-        uint16_t slotCount = 0; ///< Total slot count (0 = derive from fields.size())
-    };
-
-    /// Registers a custom product type definition to be carried through to the ConstantPool's TypeRegistry.
-    void addCustomProductType(CustomProductType def) { _customProductTypes.push_back(std::move(def)); }
-
-    /// Returns all custom product type definitions.
-    [[nodiscard]] std::vector<CustomProductType> const& customProductTypes() const
-    {
-        return _customProductTypes;
-    }
-
-    /// Describes a custom sum type (discriminated union) to be registered at target code generation time.
-    struct CustomSumType
-    {
-        std::string name;
-        std::vector<VariantInfo> variants;
-        uint16_t assignedId; ///< Pre-assigned type ID used in OALLOC instructions
-    };
-
-    /// Registers a custom sum type definition to be carried through to the ConstantPool's TypeRegistry.
-    void addCustomSumType(CustomSumType def) { _customSumTypes.push_back(std::move(def)); }
-
-    /// Returns all custom sum type definitions.
-    [[nodiscard]] std::vector<CustomSumType> const& customSumTypes() const { return _customSumTypes; }
-
-    /// Allocates the next available custom type ID (starting after builtins).
-    [[nodiscard]] uint16_t allocateCustomTypeId() { return _nextCustomTypeId++; }
-
-  private:
-    std::vector<std::pair<std::string, std::string>> _modules;
-    ConstantBoolean _trueLiteral;
-    ConstantBoolean _falseLiteral;
-    std::vector<std::unique_ptr<ConstantArray>> _constantArrays;
-    std::vector<std::unique_ptr<ConstantInt>> _numbers;
-    std::vector<std::unique_ptr<ConstantFloat>> _floats;
-    std::vector<std::unique_ptr<ConstantString>> _strings;
-    std::vector<std::unique_ptr<ConstantIP>> _ipaddrs;
-    std::vector<std::unique_ptr<ConstantCidr>> _cidrs;
-    std::vector<std::unique_ptr<ConstantRegExp>> _regexps;
-    std::vector<std::unique_ptr<IRBuiltinFunction>> _builtinFunctions;
-    std::vector<std::unique_ptr<IRFunction>> _functions;
-    std::vector<CustomProductType> _customProductTypes;
-    std::vector<CustomSumType> _customSumTypes;
-    uint16_t _nextCustomTypeId = BuiltinTypeId::LastBuiltin + 1; ///< Next type ID for custom types
-
-    friend class IRBuilder;
-};
-
-class IRBuilder
-{
-  private:
-    std::unique_ptr<IRProgram> _program;
-    IRFunction* _function;
-    BasicBlock* _insertPoint;
-    std::unordered_map<std::string, unsigned long> _nameStore;
-    SourceLocation _currentLocation;
-
-  public:
-    IRBuilder();
-    ~IRBuilder() = default;
-
-    /// Set the source location for subsequently created instructions
-    void setSourceLocation(SourceLocation loc) { _currentLocation = std::move(loc); }
-
-    /// Get the current source location
-    [[nodiscard]] SourceLocation const& sourceLocation() const noexcept { return _currentLocation; }
-
-    std::string makeName(std::string name);
-
-    void setProgram(std::unique_ptr<IRProgram> program);
-
-    IRProgram* program() const { return _program.get(); }
-
-    /// Takes ownership of the program, returning it as a unique_ptr.
-    /// After calling this, program() will return nullptr.
-    std::unique_ptr<IRProgram> takeProgram() { return std::move(_program); }
-
-    IRFunction* setFunction(IRFunction* hn);
-
-    IRFunction* function() const { return _function; }
-
-    BasicBlock* createBlock(const std::string& name);
-
-    void setInsertPoint(BasicBlock* bb);
-
-    BasicBlock* getInsertPoint() const { return _insertPoint; }
-
-    Instr* insert(std::unique_ptr<Instr> instr);
-
-    template <typename T, typename... Args>
-    T* insert(Args&&... args)
-    {
-        return static_cast<T*>(insert(std::make_unique<T>(std::forward<Args>(args)...)));
-    }
-
-    IRFunction* getFunction(const std::string& name);
-    IRFunction* findFunction(const std::string& name);
-
-    // literals
-    ConstantBoolean* getBoolean(bool literal) { return _program->getBoolean(literal); }
-
-    ConstantInt* get(int64_t literal) { return _program->get(literal); }
-
-    ConstantFloat* getFloat(double literal) { return _program->getFloat(literal); }
-
-    ConstantString* get(const std::string& literal) { return _program->get(literal); }
-
-    ConstantIP* get(const util::IPAddress& literal) { return _program->get(literal); }
-
-    ConstantCidr* get(const util::Cidr& literal) { return _program->get(literal); }
-
-    ConstantRegExp* get(const util::RegExp& literal) { return _program->get(literal); }
-
-    IRBuiltinFunction* getBuiltinFunction(const NativeCallback& cb)
-    {
-        return _program->getBuiltinFunction(cb);
-    }
-
-    ConstantArray* get(const std::vector<Constant*>& arrayElements) { return _program->get(arrayElements); }
-
-    // values
-    AllocaInstr* createAlloca(LiteralType ty, Value* arraySize, const std::string& name = "");
-    Value* createLoad(Value* value, const std::string& name = "");
-    Instr* createStore(Value* lhs, Value* rhs, const std::string& name = "");
-    Instr* createStore(Value* lhs, ConstantInt* index, Value* rhs, const std::string& name = "");
-    Instr* createPhi(const std::vector<Value*>& incomings, const std::string& name = "");
-
-    // boolean operations
-    Value* createBNot(Value* rhs, const std::string& name = ""); // !
-    Value* createBAnd(Value* lhs, Value* rhs,
-                      const std::string& name = ""); // &&
-    Value* createBOr(Value* lhs, Value* rhs,
-                     const std::string& name = ""); // ||
-    Value* createBXor(Value* lhs, Value* rhs,
-                      const std::string& name = ""); // ^
-
-    // numerical operations
-    Value* createNeg(Value* rhs, const std::string& name = "");             // -
-    Value* createNot(Value* rhs, const std::string& name = "");             // ~
-    Value* createAdd(Value* lhs, Value* rhs, const std::string& name = ""); // +
-    Value* createSub(Value* lhs, Value* rhs, const std::string& name = ""); // -
-    Value* createMul(Value* lhs, Value* rhs, const std::string& name = ""); // *
-    Value* createDiv(Value* lhs, Value* rhs, const std::string& name = ""); // /
-    Value* createRem(Value* lhs, Value* rhs, const std::string& name = ""); // %
-    Value* createShl(Value* lhs, Value* rhs, const std::string& name = ""); // <<
-    Value* createShr(Value* lhs, Value* rhs, const std::string& name = ""); // >>
-    Value* createPow(Value* lhs, Value* rhs, const std::string& name = ""); // **
-    Value* createAnd(Value* lhs, Value* rhs, const std::string& name = ""); // &
-    Value* createOr(Value* lhs, Value* rhs, const std::string& name = "");  // |
-    Value* createXor(Value* lhs, Value* rhs, const std::string& name = ""); // ^
-    Value* createNCmpEQ(Value* lhs, Value* rhs,
-                        const std::string& name = ""); // ==
-    Value* createNCmpNE(Value* lhs, Value* rhs,
-                        const std::string& name = ""); // !=
-    Value* createNCmpLE(Value* lhs, Value* rhs,
-                        const std::string& name = ""); // <=
-    Value* createNCmpGE(Value* lhs, Value* rhs,
-                        const std::string& name = ""); // >=
-    Value* createNCmpLT(Value* lhs, Value* rhs,
-                        const std::string& name = ""); // <
-    Value* createNCmpGT(Value* lhs, Value* rhs,
-                        const std::string& name = ""); // >
-
-    // float operations
-    Value* createFNeg(Value* rhs, const std::string& name = "");               // -
-    Value* createFAdd(Value* lhs, Value* rhs, const std::string& name = "");   // +
-    Value* createFSub(Value* lhs, Value* rhs, const std::string& name = "");   // -
-    Value* createFMul(Value* lhs, Value* rhs, const std::string& name = "");   // *
-    Value* createFDiv(Value* lhs, Value* rhs, const std::string& name = "");   // /
-    Value* createFRem(Value* lhs, Value* rhs, const std::string& name = "");   // %
-    Value* createFPow(Value* lhs, Value* rhs, const std::string& name = "");   // **
-    Value* createFCmpEQ(Value* lhs, Value* rhs, const std::string& name = ""); // ==
-    Value* createFCmpNE(Value* lhs, Value* rhs, const std::string& name = ""); // !=
-    Value* createFCmpLE(Value* lhs, Value* rhs, const std::string& name = ""); // <=
-    Value* createFCmpGE(Value* lhs, Value* rhs, const std::string& name = ""); // >=
-    Value* createFCmpLT(Value* lhs, Value* rhs, const std::string& name = ""); // <
-    Value* createFCmpGT(Value* lhs, Value* rhs, const std::string& name = ""); // >
-
-    // string ops
-    Value* createSAdd(Value* lhs, Value* rhs, const std::string& name = ""); // +
-    Value* createSCmpEQ(Value* lhs, Value* rhs,
-                        const std::string& name = ""); // ==
-    Value* createSCmpNE(Value* lhs, Value* rhs,
-                        const std::string& name = ""); // !=
-    Value* createSCmpLE(Value* lhs, Value* rhs,
-                        const std::string& name = ""); // <=
-    Value* createSCmpGE(Value* lhs, Value* rhs,
-                        const std::string& name = ""); // >=
-    Value* createSCmpLT(Value* lhs, Value* rhs,
-                        const std::string& name = ""); // <
-    Value* createSCmpGT(Value* lhs, Value* rhs,
-                        const std::string& name = ""); // >
-    Value* createSCmpRE(Value* lhs, Value* rhs,
-                        const std::string& name = ""); // =~
-    Value* createSCmpEB(Value* lhs, Value* rhs,
-                        const std::string& name = ""); // =^
-    Value* createSCmpEE(Value* lhs, Value* rhs,
-                        const std::string& name = "");                      // =$
-    Value* createSIn(Value* lhs, Value* rhs, const std::string& name = ""); // in
-    Value* createSLen(Value* value, const std::string& name = "");
-
-    // IP address
-    Value* createPCmpEQ(Value* lhs, Value* rhs,
-                        const std::string& name = ""); // ==
-    Value* createPCmpNE(Value* lhs, Value* rhs,
-                        const std::string& name = ""); // !=
-    Value* createPInCidr(Value* lhs, Value* rhs,
-                         const std::string& name = ""); // in
-
-    // cidr
-    // ...
-
-    // regexp
-    RegExpGroupInstr* createRegExpGroup(ConstantInt* groupId, const std::string& name = "");
-
-    // cast
-    Value* createConvert(LiteralType ty, Value* rhs,
-                         const std::string& name = ""); // cast<T>()
-    Value* createB2S(Value* rhs, const std::string& name = "");
-    Value* createN2S(Value* rhs, const std::string& name = "");
-    Value* createP2S(Value* rhs, const std::string& name = "");
-    Value* createC2S(Value* rhs, const std::string& name = "");
-    Value* createR2S(Value* rhs, const std::string& name = "");
-    Value* createS2N(Value* rhs, const std::string& name = "");
-    Value* createN2F(Value* rhs, const std::string& name = "");
-    Value* createF2N(Value* rhs, const std::string& name = "");
-    Value* createF2S(Value* rhs, const std::string& name = "");
-    Value* createS2F(Value* rhs, const std::string& name = "");
-
-    // calls
-    Instr* createCallFunction(IRBuiltinFunction* callee, std::vector<Value*> args, std::string name = "");
-    // user-defined function calls
-    FunctionCallInstr* createFunctionCall(IRFunction* callee,
-                                          std::vector<Value*> args,
-                                          const std::string& name = "",
-                                          LiteralType returnType = LiteralType::Void);
-    FunctionRetInstr* createFunctionRet(Value* result, const std::string& name = "");
-    TailCallInstr* createTailCall(IRFunction* callee, std::vector<Value*> args, const std::string& name = "");
-
-    // termination instructions
-    Instr* createRet(Value* result);
-    Instr* createBr(BasicBlock* target);
-    Instr* createCondBr(Value* condValue, BasicBlock* trueBlock, BasicBlock* falseBlock);
-    MatchInstr* createMatch(MatchClass opc, Value* cond);
-    Value* createMatchSame(Value* cond);
-    Value* createMatchHead(Value* cond);
-    Value* createMatchTail(Value* cond);
-    Value* createMatchRegExp(Value* cond);
-
-    // object operations
-    ObjAllocInstr* createObjAlloc(ConstantInt* typeId, const std::string& name = "");
-    ObjRetainInstr* createObjRetain(Value* object, const std::string& name = "");
-    ObjReleaseInstr* createObjRelease(AllocaInstr* storage, const std::string& name = "");
-    ObjGetTagInstr* createObjGetTag(Value* object, const std::string& name = "");
-    ObjSetTagInstr* createObjSetTag(Value* object, Value* tag, const std::string& name = "");
-    ObjGetSlotInstr* createObjGetSlot(Value* object, ConstantInt* slotIndex, const std::string& name = "");
-    ObjSetSlotInstr* createObjSetSlot(Value* object,
-                                      ConstantInt* slotIndex,
-                                      Value* value,
-                                      const std::string& name = "");
-    ObjTypeIdInstr* createObjTypeId(Value* object, const std::string& name = "");
-    ObjIsTypeInstr* createObjIsType(Value* object, ConstantInt* typeId, const std::string& name = "");
-
-    // Dynamic value comparison (for pattern matching with unknown types)
-    VCmpEQInstr* createVCmpEQ(Value* lhs, Value* rhs, const std::string& name = "");
-    VCmpNEInstr* createVCmpNE(Value* lhs, Value* rhs, const std::string& name = "");
-    VCmpLTInstr* createVCmpLT(Value* lhs, Value* rhs, const std::string& name = "");
-    VCmpLEInstr* createVCmpLE(Value* lhs, Value* rhs, const std::string& name = "");
-    VCmpGTInstr* createVCmpGT(Value* lhs, Value* rhs, const std::string& name = "");
-    VCmpGEInstr* createVCmpGE(Value* lhs, Value* rhs, const std::string& name = "");
-
-    // Indirect function calls (via Callable objects)
-    IndirectCallInstr* createIndirectCall(Value* callable,
-                                          std::vector<Value*> args,
-                                          const std::string& name = "");
-    IndirectTailCallInstr* createIndirectTailCall(Value* callable,
-                                                  std::vector<Value*> args,
-                                                  const std::string& name = "");
-
-    // Lazy evaluation
-    FunctionRefInstr* createFunctionRef(IRFunction* function, const std::string& name = "");
-    LazyForceInstr* createLazyForce(Value* lazyObj, const std::string& name = "");
-};
-
-/**
- * An SSA based instruction basic block.
- *
- * @see Instr, IRFunction, IRBuilder
- */
-class BasicBlock: public Value
-{
-  public:
-    BasicBlock(const std::string& name, IRFunction* parent);
-    ~BasicBlock() override;
-
-    [[nodiscard]] IRFunction* getFunction() const { return _function; }
-
-    void setParent(IRFunction* function) { _function = function; }
-
-    /*!
-     * Retrieves the last terminating instruction in this basic block.
-     *
-     * This instruction must be a termination instruction, such as
-     * a branching instruction or a function terminating instruction.
-     *
-     * @see BrInstr, CondBrInstr, MatchInstr, RetInstr
-     */
-    [[nodiscard]] TerminateInstr* getTerminator() const;
-
-    /**
-     * Checks whether this BasicBlock is assured to terminate, hence, complete.
-     *
-     * This is either achieved by having a TerminateInstr at the end or a NativeCallback
-     * that never returns.
-     */
-    [[nodiscard]] bool isComplete() const;
-
-    /**
-     * Retrieves the linear ordered list of instructions of instructions in this
-     * basic block.
-     */
-    auto instructions() { return util::unbox(_code); }
-
-    Instr* instruction(size_t i) { return _code[i].get(); }
-
-    [[nodiscard]] Instr* front() const { return _code.front().get(); }
-
-    [[nodiscard]] Instr* back() const { return _code.back().get(); }
-
-    [[nodiscard]] size_t size() const { return _code.size(); }
-
-    [[nodiscard]] bool empty() const { return _code.empty(); }
-
-    [[nodiscard]] Instr* back(size_t sub) const
-    {
-        if (sub + 1 <= _code.size())
-            return _code[_code.size() - (1 + sub)].get();
-        else
-            return nullptr;
-    }
-
-    /**
-     * Appends a new instruction, \p instr, to this basic block.
-     *
-     * The basic block will take over ownership of the given instruction.
-     */
-    Instr* push_back(std::unique_ptr<Instr> instr);
-
-    /**
-     * Inserts an instruction before the terminator (if present) or at the end.
-     * This is useful for inserting allocas into the entry block after it already has a branch.
-     */
-    Instr* insertBeforeTerminator(std::unique_ptr<Instr> instr);
-
-    /**
-     * Inserts an alloca instruction after the last existing alloca in this block.
-     * Maintains the alloca-prefix invariant: all allocas at the start of the block.
-     * This is critical for the TargetCodeGenerator's stack tracking.
-     */
-    Instr* insertAfterAllocas(std::unique_ptr<Instr> instr);
-
-    /**
-     * Removes given instruction from this basic block.
-     *
-     * The basic block will pass ownership of the given instruction to the caller.
-     * That means, the caller has to either delete \p childInstr or transfer it to
-     * another basic block.
-     *
-     * @see push_back()
-     */
-    std::unique_ptr<Instr> remove(Instr* childInstr);
-
-    /**
-     * Replaces given @p oldInstr with @p newInstr.
-     *
-     * @return returns given @p oldInstr.
-     */
-    std::unique_ptr<Instr> replace(Instr* oldInstr, std::unique_ptr<Instr> newInstr);
-
-    /**
-     * Merges given basic block's instructions into this ones end.
-     *
-     * The passed basic block's instructions <b>WILL</b> be touched, that is
-     * - all instructions will have been moved.
-     * - any successor BBs will have been relinked
-     */
-    void merge_back(BasicBlock* bb);
-
-    /**
-     * Moves this basic block after the other basic block, \p otherBB.
-     *
-     * @param otherBB the future prior basic block.
-     *
-     * In a function, all basic blocks (starting from the entry block)
-     * will be aligned linear into the execution segment.
-     *
-     * This function moves this basic block directly after
-     * the other basic block, \p otherBB.
-     *
-     * @see moveBefore()
-     */
-    void moveAfter(const BasicBlock* otherBB);
-
-    /**
-     * Moves this basic block before the other basic block, \p otherBB.
-     *
-     * @see moveAfter()
-     */
-    void moveBefore(const BasicBlock* otherBB);
-
-    /**
-     * Tests whether or not given block is straight-line located after this block.
-     *
-     * @retval true \p otherBB is straight-line located after this block.
-     * @retval false \p otherBB is not straight-line located after this block.
-     *
-     * @see moveAfter()
-     */
-    bool isAfter(const BasicBlock* otherBB) const;
-
-    /**
-     * Links given \p successor basic block to this predecessor.
-     *
-     * @param successor the basic block to link as an successor of this basic
-     *                  block.
-     *
-     * This will also automatically link this basic block as
-     * future predecessor of the \p successor.
-     *
-     * @see unlinkSuccessor()
-     * @see successors(), predecessors()
-     */
-    void linkSuccessor(BasicBlock* successor);
-
-    /**
-     * Unlink given \p successor basic block from this predecessor.
-     *
-     * @see linkSuccessor()
-     * @see successors(), predecessors()
-     */
-    void unlinkSuccessor(BasicBlock* successor);
-
-    /** Retrieves all predecessors of given basic block. */
-    [[nodiscard]] std::vector<BasicBlock*>& predecessors() { return _predecessors; }
-
-    /** Retrieves all uccessors of the given basic block. */
-    [[nodiscard]] std::vector<BasicBlock*>& successors() { return _successors; }
-
-    [[nodiscard]] const std::vector<BasicBlock*>& successors() const { return _successors; }
-
-    /** Retrieves all dominators of given basic block. */
-    [[nodiscard]] std::vector<BasicBlock*> dominators();
-
-    /** Retrieves all immediate dominators of given basic block. */
-    [[nodiscard]] std::vector<BasicBlock*> immediateDominators();
-
-    void dump();
-    [[nodiscard]] std::string dumpToString() const;
-
-    /**
-     * Performs sanity checks on internal data structures.
-     *
-     * This call does not return any success or failure as every failure is
-     * considered fatal and will cause the program to exit with diagnostics
-     * as this is most likely caused by an application programming error.
-     *
-     * @note This function is automatically invoked by IRFunction::verify()
-     *
-     * @see IRFunction::verify()
-     */
-    void verify();
-
-  private:
-    void collectIDom(std::vector<BasicBlock*>& output);
-
-  private:
-    IRFunction* _function;
-    std::vector<std::unique_ptr<Instr>> _code;
-    std::vector<BasicBlock*> _predecessors;
-    std::vector<BasicBlock*> _successors;
-
-    friend class IRBuilder;
-    friend class Instr;
-};
-
-class Runtime
-{
-  public:
-    using Value = uint64_t;
-
-    virtual ~Runtime() = default;
-
-    virtual bool import(const std::string& name,
-                        const std::string& path,
-                        std::vector<NativeCallback*>* builtins);
-
-    [[nodiscard]] NativeCallback* find(const std::string& signature) const noexcept;
-    [[nodiscard]] NativeCallback* find(const Signature& signature) const noexcept;
-
-    [[nodiscard]] auto builtins() { return util::unbox(_builtins); }
-
-    NativeCallback& registerFunction(const std::string& name);
-    NativeCallback& registerFunction(const std::string& name, LiteralType returnType);
-
-    /// Registers a builtin property with getter/setter.
-    ///
-    /// Automatically creates two NativeCallback entries (getter + setter signatures)
-    /// so the existing CALL bytecode works without VM changes.
-    ///
-    /// @param name The property name (e.g., "agent_provider").
-    /// @param type The value type of the property.
-    /// @return Reference to the NativeProperty for chaining onGet/onSet calls.
-    NativeProperty& registerProperty(std::string const& name, LiteralType type);
-
-    /// Finds a registered property by name.
-    /// @return Pointer to the property, or nullptr if not found.
-    [[nodiscard]] NativeProperty* findProperty(std::string const& name) const noexcept;
-
-    /// Returns an iterable view of all registered properties.
-    [[nodiscard]] auto const& properties() const noexcept { return _properties; }
-
-    void invoke(int id, int argc, Value* argv, Runner* cx);
-
-    /**
-     * Verifies all call instructions.
-     *
-     * @param program the IRProgram to verify.
-     * @param builder the IRBuilder to pass to each verify(), that can be used for modifications.
-     *
-     * @see bool NativeCallback::verify(Instr* call, IRBuilder* irBuilder);
-     */
-    bool verifyNativeCalls(IRProgram* program, IRBuilder* builder) const;
-
-  private:
-    std::vector<std::unique_ptr<NativeCallback>> _builtins;
-    std::vector<std::unique_ptr<NativeProperty>> _properties;
-};
-
-template <typename T, typename U>
-T* IRProgram::get(std::vector<std::unique_ptr<T>>& table, U&& literal)
-{
-    for (size_t i = 0, e = table.size(); i != e; ++i)
-        if (table[i]->get() == literal)
-            return table[i].get();
-
-    table.emplace_back(std::make_unique<T>(std::forward<U>(literal)));
-    return table.back().get();
-}
-
-template <typename T, typename U>
-T* IRProgram::get(std::vector<T>& table, U&& literal)
-{
-    if (auto i =
-            std::find_if(table.begin(), table.end(), [&](T const& elem) { return elem.get() == literal; });
-        i != table.end())
-    {
-        return &*i;
-    }
-
-    table.emplace_back(T { std::forward<U>(literal) });
-    return &table.back();
-}
-
-// {{{ inlines
-inline NativeCallback& NativeCallback::returnType(LiteralType type)
-{
-    _signature.setReturnType(type);
-    return *this;
-}
-
-template <>
-inline NativeCallback& NativeCallback::param<bool>(const std::string& name)
-{
-    _signature.args().push_back(LiteralType::Boolean);
-    _names.push_back(name);
-    _defaults.emplace_back(std::monostate {});
-
-    return *this;
-}
-
-template <>
-inline NativeCallback& NativeCallback::param<bool>(const std::string& name, bool defaultValue)
-{
-    _signature.args().push_back(LiteralType::Boolean);
-    _names.push_back(name);
-    _defaults.emplace_back(defaultValue);
-
-    return *this;
-}
-
-template <>
-inline NativeCallback& NativeCallback::param<CoreNumber>(const std::string& name)
-{
-    _signature.args().push_back(LiteralType::Number);
-    _names.push_back(name);
-    _defaults.emplace_back(std::monostate {});
-
-    return *this;
-}
-
-template <>
-inline NativeCallback& NativeCallback::param<CoreNumber>(const std::string& name, CoreNumber defaultValue)
-{
-    _signature.args().push_back(LiteralType::Number);
-    _names.push_back(name);
-    _defaults.emplace_back(defaultValue);
-
-    return *this;
-}
-
-template <>
-inline NativeCallback& NativeCallback::param<int>(const std::string& name)
-{
-    _signature.args().push_back(LiteralType::Number);
-    _names.push_back(name);
-    _defaults.emplace_back(std::monostate {});
-
-    return *this;
-}
-
-template <>
-inline NativeCallback& NativeCallback::param<int>(const std::string& name, int defaultValue)
-{
-    _signature.args().push_back(LiteralType::Number);
-    _names.push_back(name);
-    _defaults.emplace_back(CoreNumber { defaultValue });
-
-    return *this;
-}
-
-template <>
-inline NativeCallback& NativeCallback::param<CoreString>(const std::string& name)
-{
-    _signature.args().push_back(LiteralType::String);
-    _names.push_back(name);
-    _defaults.emplace_back(std::monostate {});
-
-    return *this;
-}
-
-template <>
-inline NativeCallback& NativeCallback::param<CoreString>(const std::string& name, CoreString defaultValue)
-{
-    _signature.args().push_back(LiteralType::String);
-    _names.push_back(name);
-    _defaults.emplace_back(defaultValue);
-
-    return *this;
-}
-
-template <>
-inline NativeCallback& NativeCallback::param<util::IPAddress>(const std::string& name)
-{
-    _signature.args().push_back(LiteralType::IPAddress);
-    _names.push_back(name);
-    _defaults.emplace_back(std::monostate {});
-
-    return *this;
-}
-
-template <>
-inline NativeCallback& NativeCallback::param<util::IPAddress>(const std::string& name,
-                                                              util::IPAddress defaultValue)
-{
-    _signature.args().push_back(LiteralType::IPAddress);
-    _names.push_back(name);
-    _defaults.emplace_back(defaultValue);
-
-    return *this;
-}
-
-template <>
-inline NativeCallback& NativeCallback::param<util::Cidr>(const std::string& name)
-{
-    _signature.args().push_back(LiteralType::Cidr);
-    _names.push_back(name);
-    _defaults.emplace_back(std::monostate {});
-
-    return *this;
-}
-
-template <>
-inline NativeCallback& NativeCallback::param<util::Cidr>(const std::string& name, util::Cidr defaultValue)
-{
-    _signature.args().push_back(LiteralType::Cidr);
-    _names.push_back(name);
-    _defaults.emplace_back(defaultValue);
-
-    return *this;
-}
-
-template <>
-inline NativeCallback& NativeCallback::param<util::RegExp>(const std::string& name)
-{
-    _signature.args().push_back(LiteralType::RegExp);
-    _names.push_back(name);
-    _defaults.emplace_back(std::monostate {});
-
-    return *this;
-}
-
-template <>
-inline NativeCallback& NativeCallback::param<util::RegExp>(const std::string& name, util::RegExp defaultValue)
-{
-    _signature.args().push_back(LiteralType::RegExp);
-    _names.push_back(name);
-    _defaults.emplace_back(defaultValue);
-
-    return *this;
-}
-
-template <>
-inline NativeCallback& NativeCallback::param<CoreIntArray>(const std::string& name)
-{
-    assert(_defaults.size() == _names.size());
-
-    _signature.args().push_back(LiteralType::IntArray);
-    _names.push_back(name);
-    _defaults.emplace_back(std::monostate {});
-
-    return *this;
-}
-
-template <>
-inline NativeCallback& NativeCallback::param<CoreStringArray>(const std::string& name)
-{
-    assert(_defaults.size() == _names.size());
-
-    _signature.args().push_back(LiteralType::StringArray);
-    _names.push_back(name);
-    _defaults.emplace_back(std::monostate {});
-
-    return *this;
-}
-
-template <>
-inline NativeCallback& NativeCallback::param<CoreVM::TypedObject*>(const std::string& name)
-{
-    assert(_defaults.size() == _names.size());
-
-    _signature.args().push_back(LiteralType::Object);
-    _names.push_back(name);
-    _defaults.emplace_back(std::monostate {});
-
-    return *this;
-}
-
-// ------------------------------------------------------------------------------------------
-
-template <typename... Args>
-inline NativeCallback& NativeCallback::params(Args... args)
-{
-    _signature.setArgs({ args... });
-    return *this;
-}
-
-inline NativeCallback& NativeCallback::verifier(const Verifier& vf)
-{
-    _verifier = vf;
-    return *this;
-}
-
-template <typename Class>
-inline NativeCallback& NativeCallback::verifier(bool (Class::*method)(Instr*, IRBuilder*), Class* obj)
-{
-    _verifier = std::bind(method, obj, std::placeholders::_1, std::placeholders::_2);
-    return *this;
-}
-
-template <typename Class>
-inline NativeCallback& NativeCallback::verifier(bool (Class::*method)(Instr*, IRBuilder*))
-{
-    _verifier =
-        std::bind(method, static_cast<Class*>(_runtime), std::placeholders::_1, std::placeholders::_2);
-    return *this;
-}
-
-inline bool NativeCallback::verify(Instr* call, IRBuilder* irBuilder)
-{
-    if (!_verifier)
-        return true;
-
-    return _verifier(call, irBuilder);
-}
-
-inline NativeCallback& NativeCallback::bind(const Functor& cb)
-{
-    _function = cb;
-    return *this;
-}
-
-template <typename Class>
-inline NativeCallback& NativeCallback::bind(void (Class::*method)(Params&), Class* obj)
-{
-    _function = std::bind(method, obj, std::placeholders::_1);
-    return *this;
-}
-
-template <typename Class>
-inline NativeCallback& NativeCallback::bind(void (Class::*method)(Params&))
-{
-    _function =
-        std::bind(method, static_cast<Class*>(_runtime), std::placeholders::_1, std::placeholders::_2);
-    return *this;
-}
-
-inline bool NativeCallback::parametersNamed() const
-{
-    return !_names.empty();
-}
-
-inline const std::string& NativeCallback::getParamNameAt(size_t i) const
-{
-    assert(i < _names.size());
-    return _names[i];
-}
-
-inline const NativeCallback::DefaultValue& NativeCallback::getDefaultParamAt(size_t i) const
-{
-    assert(i < _defaults.size());
-    return _defaults[i];
-}
-
-// }}}
-
-using StackPointer = size_t;
-
-class TargetCodeGenerator: public InstructionVisitor
-{
-  public:
-    TargetCodeGenerator();
-
-    std::unique_ptr<Program> generate(IRProgram* program);
-
-  protected:
-    void generate(IRFunction* function);
-
-    void dumpCurrentStack();
-
-    /**
-     * Ensures @p value is available on top of the stack.
-     *
-     * May emit a LOAD instruction if stack[sp] is not on top of the stack.
-     */
-    void emitLoad(Value* value);
-
-    void emitInstr(Opcode opc) { emitInstr(makeInstruction(opc)); }
-
-    void emitInstr(Opcode opc, Operand op1) { emitInstr(makeInstruction(opc, op1)); }
-
-    void emitInstr(Opcode opc, Operand op1, Operand op2) { emitInstr(makeInstruction(opc, op1, op2)); }
-
-    void emitInstr(Opcode opc, Operand op1, Operand op2, Operand op3)
-    {
-        emitInstr(makeInstruction(opc, op1, op2, op3));
-    }
-
-    void emitInstr(Instruction instr);
-
-    /**
-     * Emits conditional jump instruction.
-     *
-     * @param opcode Opcode for the conditional jump.
-     * @param bb Target basic block to jump to by \p opcode.
-     *
-     * This function will just emit a placeholder and will remember the
-     * instruction pointer and passed operands for later back-patching once all
-     * basic block addresses have been computed.
-     */
-    void emitCondJump(Opcode opcode, BasicBlock* bb);
-
-    /**
-     * Emits unconditional jump instruction.
-     *
-     * @param bb Target basic block to jump to by \p opcode.
-     *
-     * This function will just emit a placeholder and will remember the
-     * instruction pointer and passed operands for later back-patching once all
-     * basic block addresses have been computed.
-     */
-    void emitJump(BasicBlock* bb);
-
-    void emitBinaryAssoc(Instr& instr, Opcode opcode);
-    void emitBinary(Instr& instr, Opcode opcode);
-    void emitUnary(Instr& instr, Opcode opcode);
-
-    Operand getConstantInt(Value* value);
-
-    /**
-     * Retrieves the instruction pointer of the next instruction to be emitted.
-     */
-    size_t getInstructionPointer() const { return _code.size(); }
-
-    /**
-     * Finds given variable on global storage and returns its absolute offset if found or -1 if not.
-     */
-    std::optional<size_t> findGlobal(const Value* variable) const;
-
-    /**
-     * Retrieves the current number of elements on the stack.
-     *
-     * This also means, this value will be the absolute index for the next value
-     * to be placed on top of the stack.
-     */
-    StackPointer getStackPointer() const { return _stack.size(); }
-
-    /** Locates given @p value on the stack.
-     */
-    StackPointer getStackPointer(const Value* value);
-
-    void changeStack(size_t pops, const Value* pushValue);
-    void pop(size_t count);
-    void push(const Value* alias);
-
-    void visit(NopInstr& instr) override;
-
-    // storage
-    void visit(AllocaInstr& instr) override;
-    void visit(StoreInstr& instr) override;
-    void visit(LoadInstr& instr) override;
-    void visit(PhiNode& instr) override;
-
-    // calls
-    void visit(CallInstr& instr) override;
-
-    void visit(FunctionCallInstr& instr) override;
-    void visit(FunctionRetInstr& instr) override;
-    void visit(TailCallInstr& instr) override;
-    void visit(IndirectCallInstr& instr) override;
-    void visit(IndirectTailCallInstr& instr) override;
-
-    // terminator
-    void visit(CondBrInstr& instr) override;
-    void visit(BrInstr& instr) override;
-    void visit(RetInstr& instr) override;
-    void visit(MatchInstr& instr) override;
-
-    // regexp
-    void visit(RegExpGroupInstr& instr) override;
-
-    // type cast
-    void visit(CastInstr& instr) override;
-
-    // numeric
-    void visit(INegInstr& instr) override;
-    void visit(INotInstr& instr) override;
-    void visit(IAddInstr& instr) override;
-    void visit(ISubInstr& instr) override;
-    void visit(IMulInstr& instr) override;
-    void visit(IDivInstr& instr) override;
-    void visit(IRemInstr& instr) override;
-    void visit(IPowInstr& instr) override;
-    void visit(IAndInstr& instr) override;
-    void visit(IOrInstr& instr) override;
-    void visit(IXorInstr& instr) override;
-    void visit(IShlInstr& instr) override;
-    void visit(IShrInstr& instr) override;
-    void visit(ICmpEQInstr& instr) override;
-    void visit(ICmpNEInstr& instr) override;
-    void visit(ICmpLEInstr& instr) override;
-    void visit(ICmpGEInstr& instr) override;
-    void visit(ICmpLTInstr& instr) override;
-    void visit(ICmpGTInstr& instr) override;
-
-    // boolean
-    void visit(BNotInstr& instr) override;
-    void visit(BAndInstr& instr) override;
-    void visit(BOrInstr& instr) override;
-    void visit(BXorInstr& instr) override;
-
-    // string
-    void visit(SLenInstr& instr) override;
-    void visit(SIsEmptyInstr& instr) override;
-    void visit(SAddInstr& instr) override;
-    void visit(SSubStrInstr& instr) override;
-    void visit(SCmpEQInstr& instr) override;
-    void visit(SCmpNEInstr& instr) override;
-    void visit(SCmpLEInstr& instr) override;
-    void visit(SCmpGEInstr& instr) override;
-    void visit(SCmpLTInstr& instr) override;
-    void visit(SCmpGTInstr& instr) override;
-    void visit(SCmpREInstr& instr) override;
-    void visit(SCmpBegInstr& instr) override;
-    void visit(SCmpEndInstr& instr) override;
-    void visit(SInInstr& instr) override;
-
-    // ip
-    void visit(PCmpEQInstr& instr) override;
-    void visit(PCmpNEInstr& instr) override;
-    void visit(PInCidrInstr& instr) override;
-
-    // object operations
-    void visit(ObjAllocInstr& instr) override;
-    void visit(ObjRetainInstr& instr) override;
-    void visit(ObjReleaseInstr& instr) override;
-    void visit(ObjGetTagInstr& instr) override;
-    void visit(ObjSetTagInstr& instr) override;
-    void visit(ObjGetSlotInstr& instr) override;
-    void visit(ObjSetSlotInstr& instr) override;
-    void visit(ObjTypeIdInstr& instr) override;
-    void visit(ObjIsTypeInstr& instr) override;
-    void visit(VCmpEQInstr& instr) override;
-    void visit(VCmpNEInstr& instr) override;
-    void visit(VCmpLTInstr& instr) override;
-    void visit(VCmpLEInstr& instr) override;
-    void visit(VCmpGTInstr& instr) override;
-    void visit(VCmpGEInstr& instr) override;
-
-    // float
-    void visit(FNegInstr& instr) override;
-    void visit(FAddInstr& instr) override;
-    void visit(FSubInstr& instr) override;
-    void visit(FMulInstr& instr) override;
-    void visit(FDivInstr& instr) override;
-    void visit(FRemInstr& instr) override;
-    void visit(FPowInstr& instr) override;
-    void visit(FCmpEQInstr& instr) override;
-    void visit(FCmpNEInstr& instr) override;
-    void visit(FCmpLEInstr& instr) override;
-    void visit(FCmpGEInstr& instr) override;
-    void visit(FCmpLTInstr& instr) override;
-    void visit(FCmpGTInstr& instr) override;
-
-    // lazy evaluation
-    void visit(FunctionRefInstr& instr) override;
-    void visit(LazyForceInstr& instr) override;
-
-  private:
-    struct ConditionalJump
-    {
-        size_t pc;
-        Opcode opcode;
-    };
-
-    struct UnconditionalJump
-    {
-        size_t pc;
-        Opcode opcode;
-    };
-
-    //! list of raised errors during code generation.
-    std::vector<std::string> _errors;
-
-    std::unordered_map<BasicBlock*, std::list<ConditionalJump>> _conditionalJumps;
-    std::unordered_map<BasicBlock*, std::list<UnconditionalJump>> _unconditionalJumps;
-    std::list<std::pair<MatchInstr*, size_t /*matchId*/>> _matchHints;
-
-    size_t _functionId;             //!< current function's ID
-    std::vector<Instruction> _code; //!< current function's code
-
-    /** Location table for current function (sparse: only records when location changes) */
-    ConstantPool::LocationTable _locationTable;
-    SourceLocation _lastRecordedLocation;
-
-    /** target stack during target code generation */
-    std::deque<const Value*> _stack;
-
-    /** Maps alloca instructions to their fixed stack indices.
-     *  Allocas are assigned indices sequentially as they're encountered.
-     */
-    std::unordered_map<const Value*, size_t> _allocaIndices;
-
-    /** Number of allocas seen so far (also the next alloca index) */
-    size_t _allocaCount = 0;
-
-    /** Parameter allocas that are already on the stack from UCALL.
-     *  These must not emit ALLOC opcodes or be pushed to _stack again. */
-    std::unordered_set<const AllocaInstr*> _parameterAllocas;
-
-    /** global scope mapping */
-    std::deque<const Value*> _globals;
-
-    // target program output
-    ConstantPool _cp;
-};
-
-/** Implements SMATCHEQ instruction. */
-class MatchSame: public Match
-{
-  public:
-    MatchSame(const MatchDef& def, Program* program);
-
-    uint64_t evaluate(const CoreString* condition, Runner* env) const override;
-
-  private:
-    std::unordered_map<CoreString, uint64_t> _map;
-};
-
-/** Implements SMATCHBEG instruction. */
-class MatchHead: public Match
-{
-  public:
-    MatchHead(const MatchDef& def, Program* program);
-
-    uint64_t evaluate(const CoreString* condition, Runner* env) const override;
-
-  private:
-    util::PrefixTree<CoreString, uint64_t> _map;
-};
-
-/** Implements SMATCHBEG instruction. */
-class MatchTail: public Match
-{
-  public:
-    MatchTail(const MatchDef& def, Program* program);
-
-    uint64_t evaluate(const CoreString* condition, Runner* env) const override;
-
-  private:
-    util::SuffixTree<CoreString, uint64_t> _map;
-};
-
-/** Implements SMATCHR instruction. */
-class MatchRegEx: public Match
-{
-  public:
-    MatchRegEx(const MatchDef& def, Program* program);
-
-    uint64_t evaluate(const CoreString* condition, Runner* env) const override;
-
-  private:
-    std::vector<std::pair<util::RegExp, uint64_t>> _map;
-};
-
-} // namespace CoreVM
-
-namespace CoreVM::transform
-{
-
-/**
- * Eliminates empty blocks, that are just jumping to the next block.
- */
-bool emptyBlockElimination(IRFunction* function);
-bool rewriteCondBrToSameBranches(IRFunction* function);
-bool eliminateUnusedInstr(IRFunction* function);
-bool eliminateLinearBr(IRFunction* function);
-bool foldConstantCondBr(IRFunction* function);
-bool rewriteBrToExit(IRFunction* function);
-
-/**
- * Merges equal blocks into one, eliminating duplicated blocks.
- *
- * A block is equal if their instructions and their successors are equal.
- */
-bool mergeSameBlocks(IRFunction* function);
-
-/**
- * Eliminates empty blocks, that are just jumping to the next block.
- */
-bool eliminateUnusedBlocks(IRFunction* function);
-
-} // namespace CoreVM::transform
-
-namespace CoreVM::diagnostics
-{
-
-enum class Type
-{
-    TokenError,
-    SyntaxError,
-    TypeError,
-    Warning,
-    LinkError
-};
-
-/// Diagnostic message with optional suggestions and source context.
-struct Message
-{
-    Type type;
-    SourceLocation sourceLocation;
-    std::string text;
-    std::vector<std::string> suggestions;      ///< Optional hints for fixing the error
-    std::optional<std::string> contextSnippet; ///< Source line for display context
-
-    /// Constructs a message without suggestions (backward compatible).
-    Message(Type ty, SourceLocation sl, std::string t):
-        type { ty }, sourceLocation { std::move(sl) }, text { std::move(t) }
-    {
-    }
-
-    /// Constructs a message with suggestions and optional context snippet.
-    Message(Type ty,
-            SourceLocation sl,
-            std::string t,
-            std::vector<std::string> sugg,
-            std::optional<std::string> ctx = std::nullopt):
-        type { ty },
-        sourceLocation { std::move(sl) },
-        text { std::move(t) },
-        suggestions { std::move(sugg) },
-        contextSnippet { std::move(ctx) }
-    {
-    }
-
-    [[nodiscard]] std::string string() const;
-
-    bool operator==(const Message& other) const noexcept;
-
-    bool operator!=(const Message& other) const noexcept { return !(*this == other); }
-};
-
-class Report
-{
-  public:
-    virtual ~Report() = default;
-
-    template <typename... Args>
-    void tokenError(const SourceLocation& sloc, std::format_string<Args...> f, Args... args)
-    {
-        emplace_back(Type::TokenError, sloc, std::format(f, std::move(args)...));
-    }
-
-    template <typename... Args>
-    void syntaxError(const SourceLocation& sloc, std::format_string<Args...> f, Args... args)
-    {
-        emplace_back(Type::SyntaxError, sloc, std::format(f, std::move(args)...));
-    }
-
-    /// Reports a syntax error with suggestions for fixing.
-    template <typename... Args>
-    void syntaxErrorWithSuggestions(SourceLocation const& sloc,
-                                    std::vector<std::string> suggestions,
-                                    std::optional<std::string> contextSnippet,
-                                    std::format_string<Args...> f,
-                                    Args... args)
-    {
-        push_back(Message(Type::SyntaxError,
-                          sloc,
-                          std::format(f, std::move(args)...),
-                          std::move(suggestions),
-                          std::move(contextSnippet)));
-    }
-
-    template <typename... Args>
-    void typeError(const SourceLocation& sloc, std::format_string<Args...> f, Args... args)
-    {
-        emplace_back(Type::TypeError, sloc, std::format(f, std::move(args)...));
-    }
-
-    /// Reports a type error with suggestions for fixing.
-    template <typename... Args>
-    void typeErrorWithSuggestions(SourceLocation const& sloc,
-                                  std::vector<std::string> suggestions,
-                                  std::optional<std::string> contextSnippet,
-                                  std::format_string<Args...> f,
-                                  Args... args)
-    {
-        push_back(Message(Type::TypeError,
-                          sloc,
-                          std::format(f, std::move(args)...),
-                          std::move(suggestions),
-                          std::move(contextSnippet)));
-    }
-
-    template <typename... Args>
-    void warning(const SourceLocation& sloc, std::format_string<Args...> f, Args... args)
-    {
-        emplace_back(Type::Warning, sloc, std::format(f, std::move(args)...));
-    }
-
-    template <typename... Args>
-    void linkError(std::format_string<Args...> f, Args... args)
-    {
-        emplace_back(
-            Type::LinkError, SourceLocation {}, std::vformat(f.get(), std::make_format_args(args...)));
-    }
-
-    void emplace_back(Type ty, SourceLocation sl, std::string t)
-    {
-        push_back(Message(ty, std::move(sl), std::move(t)));
-    }
-
-    virtual void push_back(Message msg) = 0;
-    [[nodiscard]] virtual bool containsFailures() const noexcept = 0;
-};
-
-using MessageList = std::vector<Message>;
-
-class BufferedReport: public Report
-{
-  public:
-    void push_back(Message msg) override;
-    [[nodiscard]] bool containsFailures() const noexcept override;
-
-    void log() const;
-
-    [[nodiscard]] const MessageList& messages() const noexcept { return _messages; }
-
-    void clear();
-
-    [[nodiscard]] size_t size() const noexcept { return _messages.size(); }
-
-    /// Truncates the message list to the given size, discarding newer messages.
-    void truncate(size_t n)
-    {
-        if (n < _messages.size())
-            _messages.erase(_messages.begin() + static_cast<ptrdiff_t>(n), _messages.end());
-    }
-
-    const Message& operator[](size_t i) const { return _messages[i]; }
-
-    using iterator = MessageList::iterator;
-    using const_iterator = MessageList::const_iterator;
-
-    iterator begin() noexcept { return _messages.begin(); }
-
-    iterator end() noexcept { return _messages.end(); }
-
-    [[nodiscard]] const_iterator begin() const noexcept { return _messages.begin(); }
-
-    [[nodiscard]] const_iterator end() const noexcept { return _messages.end(); }
-
-    [[nodiscard]] bool contains(const Message& m) const noexcept;
-
-    bool operator==(const BufferedReport& other) const noexcept;
-
-    bool operator!=(const BufferedReport& other) const noexcept { return !(*this == other); }
-
-  private:
-    MessageList _messages;
-};
-
-class ConsoleReport: public Report
-{
-  public:
-    ConsoleReport();
-
-    void push_back(Message msg) override;
-    [[nodiscard]] bool containsFailures() const noexcept override;
-
-  private:
-    size_t _errorCount;
-};
-
-std::ostream& operator<<(std::ostream& os, const Report& report);
-
-using DifferenceReport = std::pair<MessageList, MessageList>;
-
-DifferenceReport difference(const BufferedReport& first, const BufferedReport& second);
-
-class DiagnosticsError: public std::runtime_error
-{
-  public:
-    explicit DiagnosticsError(SourceLocation sloc, const std::string& msg):
-        std::runtime_error { msg }, _sloc { std::move(sloc) }
-    {
-    }
-
-    [[nodiscard]] const SourceLocation& sourceLocation() const noexcept { return _sloc; }
-
-  private:
-    SourceLocation _sloc;
-};
-
-class LexerError: public DiagnosticsError
-{
-  public:
-    LexerError(SourceLocation sloc, const std::string& msg): DiagnosticsError { std::move(sloc), msg } {}
-};
-
-class SyntaxError: public DiagnosticsError
-{
-  public:
-    SyntaxError(SourceLocation sloc, const std::string& msg): DiagnosticsError { std::move(sloc), msg } {}
-};
-
-class TypeError: public DiagnosticsError
-{
-  public:
-    TypeError(SourceLocation sloc, const std::string& msg): DiagnosticsError { std::move(sloc), msg } {}
-};
-
-} // namespace CoreVM::diagnostics
-
-namespace std
-{
-
-template <>
-struct hash<CoreVM::util::Cidr>
-{
-    size_t operator()(const CoreVM::util::Cidr& v) const noexcept
-    {
-        auto const* bytes = static_cast<const uint8_t*>(v.address().data());
-        auto const len = v.address().size();
-        size_t h = 0;
-        for (size_t i = 0; i < len; ++i)
-            h = h * 131 + bytes[i];
-        return h ^ static_cast<size_t>(v.prefix());
-    }
-};
-
-template <>
-struct hash<CoreVM::LiteralType>
-{
-    uint32_t operator()(CoreVM::LiteralType v) const noexcept { return static_cast<uint32_t>(v); }
-};
-
-} // namespace std
-
-template <>
-struct std::formatter<CoreVM::util::Cidr>: std::formatter<std::string>
-{
-    auto format(CoreVM::util::Cidr const& value, std::format_context& ctx) const
-        -> std::format_context::iterator
-    {
-        return std::formatter<std::string>::format(value.str(), ctx);
-    }
-};
-
-template <>
-struct std::formatter<CoreVM::Signature>: std::formatter<std::string>
-{
-    auto format(const CoreVM::Signature& v, std::format_context& ctx) const -> std::format_context::iterator
-    {
-        return std::formatter<std::string>::format(v.to_s(), ctx);
-    }
-};
-
-template <>
-struct std::formatter<CoreVM::util::RegExp>: std::formatter<std::string>
-{
-    auto format(CoreVM::util::RegExp const& v, std::format_context& ctx) const
-        -> std::format_context::iterator
-    {
-        return std::formatter<std::string>::format(v.pattern(), ctx);
-    }
-};
-
-template <>
-struct std::formatter<CoreVM::diagnostics::Type>: std::formatter<std::string_view>
-{
-    auto format(const CoreVM::diagnostics::Type& value, std::format_context& ctx) const
-        -> std::format_context::iterator
-    {
-        std::string_view name;
-        switch (value)
-        {
-            case CoreVM::diagnostics::Type::TokenError: name = "TokenError"; break;
-            case CoreVM::diagnostics::Type::SyntaxError: name = "SyntaxError"; break;
-            case CoreVM::diagnostics::Type::TypeError: name = "TypeError"; break;
-            case CoreVM::diagnostics::Type::Warning: name = "Warning"; break;
-            case CoreVM::diagnostics::Type::LinkError: name = "LinkError"; break;
-        }
-        return std::formatter<std::string_view>::format(name, ctx);
-    }
-};
-
-template <>
-struct std::formatter<CoreVM::diagnostics::Message>: std::formatter<std::string>
-{
-    auto format(const CoreVM::diagnostics::Message& value, std::format_context& ctx) const
-        -> std::format_context::iterator
-    {
-        return std::formatter<std::string>::format(value.string(), ctx);
-    }
-};
-
-template <>
-struct std::formatter<CoreVM::LiteralType>: std::formatter<std::string_view>
-{
-    auto format(CoreVM::LiteralType id, std::format_context& ctx) const -> std::format_context::iterator
-    {
-        std::string_view name;
-        switch (id)
-        {
-            case CoreVM::LiteralType::Void: name = "Void"; break;
-            case CoreVM::LiteralType::Boolean: name = "Boolean"; break;
-            case CoreVM::LiteralType::Number: name = "Number"; break;
-            case CoreVM::LiteralType::String: name = "String"; break;
-            case CoreVM::LiteralType::IPAddress: name = "IPAddress"; break;
-            case CoreVM::LiteralType::Cidr: name = "Cidr"; break;
-            case CoreVM::LiteralType::RegExp: name = "RegExp"; break;
-            case CoreVM::LiteralType::Function: name = "Function"; break;
-            case CoreVM::LiteralType::IntArray: name = "IntArray"; break;
-            case CoreVM::LiteralType::StringArray: name = "StringArray"; break;
-            case CoreVM::LiteralType::IPAddrArray: name = "IPAddrArray"; break;
-            case CoreVM::LiteralType::CidrArray: name = "CidrArray"; break;
-            case CoreVM::LiteralType::IntPair: name = "IntPair"; break;
-            case CoreVM::LiteralType::Option: name = "Option"; break;
-            case CoreVM::LiteralType::Result: name = "Result"; break;
-            case CoreVM::LiteralType::Object: name = "Object"; break;
-            case CoreVM::LiteralType::Float: name = "Float"; break;
-        }
-        return std::formatter<std::string_view>::format(name, ctx);
-    }
-};
-
-template <>
-struct std::formatter<CoreVM::FilePos>: std::formatter<std::string>
-{
-    auto format(const CoreVM::FilePos& value, std::format_context& ctx) const -> std::format_context::iterator
-    {
-        return std::formatter<std::string>::format(CoreVM::tos(value), ctx);
-    }
-};
-
-template <>
-struct std::formatter<CoreVM::SourceLocation>: std::formatter<std::string>
-{
-    auto format(const CoreVM::SourceLocation& value, std::format_context& ctx) const
-        -> std::format_context::iterator
-    {
-        if (!value.filename.empty())
-            return std::formatter<std::string>::format(value.filename + ":" + CoreVM::tos(value.begin), ctx);
-        else
-            return std::formatter<std::string>::format(CoreVM::tos(value.begin), ctx);
-    }
-};
+#include <CoreVM/vm/NativeCallback.hpp>
+#include <CoreVM/vm/Program.hpp>
+#include <CoreVM/vm/Runner.hpp>
+#include <CoreVM/vm/Runtime.hpp>
