+++ src/CoreVM/vm/Runner.cpp
@@ -12,6 +12,7 @@
 #include <cstdlib>
 #include <cstring>
 #include <memory>
+#include <print>
 #include <ranges>
 #include <utility>
 #include <vector>
